[tensorflow-sys] Expose the eager API.
