+++ tensorflow-sys/Cargo.toml
@@ -31,6 +31,7 @@ zip = "0.5.13"
 
 [features]
 tensorflow_gpu = []
+eager = []
 # This is for testing purposes; users should not use this.
 examples_system_alloc = []
 private-docs-rs = [] # DO NOT RELY ON THIS
