--- tensorflow-sys/src/bindgen.rs
@@ -1,4813 +1,0 @@
-/* automatically generated by rust-bindgen 0.55.1 */
-
-pub const _STDINT_H: u32 = 1;
-pub const _FEATURES_H: u32 = 1;
-pub const _DEFAULT_SOURCE: u32 = 1;
-pub const __GLIBC_USE_ISOC2X: u32 = 0;
-pub const __USE_ISOC11: u32 = 1;
-pub const __USE_ISOC99: u32 = 1;
-pub const __USE_ISOC95: u32 = 1;
-pub const __USE_POSIX_IMPLICITLY: u32 = 1;
-pub const _POSIX_SOURCE: u32 = 1;
-pub const _POSIX_C_SOURCE: u32 = 200809;
-pub const __USE_POSIX: u32 = 1;
-pub const __USE_POSIX2: u32 = 1;
-pub const __USE_POSIX199309: u32 = 1;
-pub const __USE_POSIX199506: u32 = 1;
-pub const __USE_XOPEN2K: u32 = 1;
-pub const __USE_XOPEN2K8: u32 = 1;
-pub const _ATFILE_SOURCE: u32 = 1;
-pub const __USE_MISC: u32 = 1;
-pub const __USE_ATFILE: u32 = 1;
-pub const __USE_FORTIFY_LEVEL: u32 = 0;
-pub const __GLIBC_USE_DEPRECATED_GETS: u32 = 0;
-pub const __GLIBC_USE_DEPRECATED_SCANF: u32 = 0;
-pub const _STDC_PREDEF_H: u32 = 1;
-pub const __STDC_IEC_559__: u32 = 1;
-pub const __STDC_IEC_559_COMPLEX__: u32 = 1;
-pub const __STDC_ISO_10646__: u32 = 201706;
-pub const __GNU_LIBRARY__: u32 = 6;
-pub const __GLIBC__: u32 = 2;
-pub const __GLIBC_MINOR__: u32 = 31;
-pub const _SYS_CDEFS_H: u32 = 1;
-pub const __glibc_c99_flexarr_available: u32 = 1;
-pub const __WORDSIZE: u32 = 64;
-pub const __WORDSIZE_TIME64_COMPAT32: u32 = 1;
-pub const __SYSCALL_WORDSIZE: u32 = 64;
-pub const __LONG_DOUBLE_USES_FLOAT128: u32 = 0;
-pub const __HAVE_GENERIC_SELECTION: u32 = 1;
-pub const __GLIBC_USE_LIB_EXT2: u32 = 0;
-pub const __GLIBC_USE_IEC_60559_BFP_EXT: u32 = 0;
-pub const __GLIBC_USE_IEC_60559_BFP_EXT_C2X: u32 = 0;
-pub const __GLIBC_USE_IEC_60559_FUNCS_EXT: u32 = 0;
-pub const __GLIBC_USE_IEC_60559_FUNCS_EXT_C2X: u32 = 0;
-pub const __GLIBC_USE_IEC_60559_TYPES_EXT: u32 = 0;
-pub const _BITS_TYPES_H: u32 = 1;
-pub const __TIMESIZE: u32 = 64;
-pub const _BITS_TYPESIZES_H: u32 = 1;
-pub const __OFF_T_MATCHES_OFF64_T: u32 = 1;
-pub const __INO_T_MATCHES_INO64_T: u32 = 1;
-pub const __RLIM_T_MATCHES_RLIM64_T: u32 = 1;
-pub const __STATFS_MATCHES_STATFS64: u32 = 1;
-pub const __FD_SETSIZE: u32 = 1024;
-pub const _BITS_TIME64_H: u32 = 1;
-pub const _BITS_WCHAR_H: u32 = 1;
-pub const _BITS_STDINT_INTN_H: u32 = 1;
-pub const _BITS_STDINT_UINTN_H: u32 = 1;
-pub const INT8_MIN: i32 = -128;
-pub const INT16_MIN: i32 = -32768;
-pub const INT32_MIN: i32 = -2147483648;
-pub const INT8_MAX: u32 = 127;
-pub const INT16_MAX: u32 = 32767;
-pub const INT32_MAX: u32 = 2147483647;
-pub const UINT8_MAX: u32 = 255;
-pub const UINT16_MAX: u32 = 65535;
-pub const UINT32_MAX: u32 = 4294967295;
-pub const INT_LEAST8_MIN: i32 = -128;
-pub const INT_LEAST16_MIN: i32 = -32768;
-pub const INT_LEAST32_MIN: i32 = -2147483648;
-pub const INT_LEAST8_MAX: u32 = 127;
-pub const INT_LEAST16_MAX: u32 = 32767;
-pub const INT_LEAST32_MAX: u32 = 2147483647;
-pub const UINT_LEAST8_MAX: u32 = 255;
-pub const UINT_LEAST16_MAX: u32 = 65535;
-pub const UINT_LEAST32_MAX: u32 = 4294967295;
-pub const INT_FAST8_MIN: i32 = -128;
-pub const INT_FAST16_MIN: i64 = -9223372036854775808;
-pub const INT_FAST32_MIN: i64 = -9223372036854775808;
-pub const INT_FAST8_MAX: u32 = 127;
-pub const INT_FAST16_MAX: u64 = 9223372036854775807;
-pub const INT_FAST32_MAX: u64 = 9223372036854775807;
-pub const UINT_FAST8_MAX: u32 = 255;
-pub const UINT_FAST16_MAX: i32 = -1;
-pub const UINT_FAST32_MAX: i32 = -1;
-pub const INTPTR_MIN: i64 = -9223372036854775808;
-pub const INTPTR_MAX: u64 = 9223372036854775807;
-pub const UINTPTR_MAX: i32 = -1;
-pub const PTRDIFF_MIN: i64 = -9223372036854775808;
-pub const PTRDIFF_MAX: u64 = 9223372036854775807;
-pub const SIG_ATOMIC_MIN: i32 = -2147483648;
-pub const SIG_ATOMIC_MAX: u32 = 2147483647;
-pub const SIZE_MAX: i32 = -1;
-pub const WINT_MIN: u32 = 0;
-pub const WINT_MAX: u32 = 4294967295;
-pub const true_: u32 = 1;
-pub const false_: u32 = 0;
-pub const __bool_true_false_are_defined: u32 = 1;
-pub const _STDLIB_H: u32 = 1;
-pub const WNOHANG: u32 = 1;
-pub const WUNTRACED: u32 = 2;
-pub const WSTOPPED: u32 = 2;
-pub const WEXITED: u32 = 4;
-pub const WCONTINUED: u32 = 8;
-pub const WNOWAIT: u32 = 16777216;
-pub const __WNOTHREAD: u32 = 536870912;
-pub const __WALL: u32 = 1073741824;
-pub const __WCLONE: u32 = 2147483648;
-pub const __ENUM_IDTYPE_T: u32 = 1;
-pub const __W_CONTINUED: u32 = 65535;
-pub const __WCOREFLAG: u32 = 128;
-pub const __HAVE_FLOAT128: u32 = 0;
-pub const __HAVE_DISTINCT_FLOAT128: u32 = 0;
-pub const __HAVE_FLOAT64X: u32 = 1;
-pub const __HAVE_FLOAT64X_LONG_DOUBLE: u32 = 1;
-pub const __HAVE_FLOAT16: u32 = 0;
-pub const __HAVE_FLOAT32: u32 = 1;
-pub const __HAVE_FLOAT64: u32 = 1;
-pub const __HAVE_FLOAT32X: u32 = 1;
-pub const __HAVE_FLOAT128X: u32 = 0;
-pub const __HAVE_DISTINCT_FLOAT16: u32 = 0;
-pub const __HAVE_DISTINCT_FLOAT32: u32 = 0;
-pub const __HAVE_DISTINCT_FLOAT64: u32 = 0;
-pub const __HAVE_DISTINCT_FLOAT32X: u32 = 0;
-pub const __HAVE_DISTINCT_FLOAT64X: u32 = 0;
-pub const __HAVE_DISTINCT_FLOAT128X: u32 = 0;
-pub const __HAVE_FLOATN_NOT_TYPEDEF: u32 = 0;
-pub const __ldiv_t_defined: u32 = 1;
-pub const __lldiv_t_defined: u32 = 1;
-pub const RAND_MAX: u32 = 2147483647;
-pub const EXIT_FAILURE: u32 = 1;
-pub const EXIT_SUCCESS: u32 = 0;
-pub const _SYS_TYPES_H: u32 = 1;
-pub const __clock_t_defined: u32 = 1;
-pub const __clockid_t_defined: u32 = 1;
-pub const __time_t_defined: u32 = 1;
-pub const __timer_t_defined: u32 = 1;
-pub const __BIT_TYPES_DEFINED__: u32 = 1;
-pub const _ENDIAN_H: u32 = 1;
-pub const _BITS_ENDIAN_H: u32 = 1;
-pub const __LITTLE_ENDIAN: u32 = 1234;
-pub const __BIG_ENDIAN: u32 = 4321;
-pub const __PDP_ENDIAN: u32 = 3412;
-pub const _BITS_ENDIANNESS_H: u32 = 1;
-pub const __BYTE_ORDER: u32 = 1234;
-pub const __FLOAT_WORD_ORDER: u32 = 1234;
-pub const LITTLE_ENDIAN: u32 = 1234;
-pub const BIG_ENDIAN: u32 = 4321;
-pub const PDP_ENDIAN: u32 = 3412;
-pub const BYTE_ORDER: u32 = 1234;
-pub const _BITS_BYTESWAP_H: u32 = 1;
-pub const _BITS_UINTN_IDENTITY_H: u32 = 1;
-pub const _SYS_SELECT_H: u32 = 1;
-pub const __FD_ZERO_STOS: &'static [u8; 6usize] = b"stosq\0";
-pub const __sigset_t_defined: u32 = 1;
-pub const __timeval_defined: u32 = 1;
-pub const _STRUCT_TIMESPEC: u32 = 1;
-pub const FD_SETSIZE: u32 = 1024;
-pub const _BITS_PTHREADTYPES_COMMON_H: u32 = 1;
-pub const _THREAD_SHARED_TYPES_H: u32 = 1;
-pub const _BITS_PTHREADTYPES_ARCH_H: u32 = 1;
-pub const __SIZEOF_PTHREAD_MUTEX_T: u32 = 40;
-pub const __SIZEOF_PTHREAD_ATTR_T: u32 = 56;
-pub const __SIZEOF_PTHREAD_RWLOCK_T: u32 = 56;
-pub const __SIZEOF_PTHREAD_BARRIER_T: u32 = 32;
-pub const __SIZEOF_PTHREAD_MUTEXATTR_T: u32 = 4;
-pub const __SIZEOF_PTHREAD_COND_T: u32 = 48;
-pub const __SIZEOF_PTHREAD_CONDATTR_T: u32 = 4;
-pub const __SIZEOF_PTHREAD_RWLOCKATTR_T: u32 = 8;
-pub const __SIZEOF_PTHREAD_BARRIERATTR_T: u32 = 4;
-pub const _THREAD_MUTEX_INTERNAL_H: u32 = 1;
-pub const __PTHREAD_MUTEX_HAVE_PREV: u32 = 1;
-pub const __have_pthread_attr_t: u32 = 1;
-pub const _ALLOCA_H: u32 = 1;
-pub const _LIBC_LIMITS_H_: u32 = 1;
-pub const MB_LEN_MAX: u32 = 16;
-pub const _BITS_POSIX1_LIM_H: u32 = 1;
-pub const _POSIX_AIO_LISTIO_MAX: u32 = 2;
-pub const _POSIX_AIO_MAX: u32 = 1;
-pub const _POSIX_ARG_MAX: u32 = 4096;
-pub const _POSIX_CHILD_MAX: u32 = 25;
-pub const _POSIX_DELAYTIMER_MAX: u32 = 32;
-pub const _POSIX_HOST_NAME_MAX: u32 = 255;
-pub const _POSIX_LINK_MAX: u32 = 8;
-pub const _POSIX_LOGIN_NAME_MAX: u32 = 9;
-pub const _POSIX_MAX_CANON: u32 = 255;
-pub const _POSIX_MAX_INPUT: u32 = 255;
-pub const _POSIX_MQ_OPEN_MAX: u32 = 8;
-pub const _POSIX_MQ_PRIO_MAX: u32 = 32;
-pub const _POSIX_NAME_MAX: u32 = 14;
-pub const _POSIX_NGROUPS_MAX: u32 = 8;
-pub const _POSIX_OPEN_MAX: u32 = 20;
-pub const _POSIX_PATH_MAX: u32 = 256;
-pub const _POSIX_PIPE_BUF: u32 = 512;
-pub const _POSIX_RE_DUP_MAX: u32 = 255;
-pub const _POSIX_RTSIG_MAX: u32 = 8;
-pub const _POSIX_SEM_NSEMS_MAX: u32 = 256;
-pub const _POSIX_SEM_VALUE_MAX: u32 = 32767;
-pub const _POSIX_SIGQUEUE_MAX: u32 = 32;
-pub const _POSIX_SSIZE_MAX: u32 = 32767;
-pub const _POSIX_STREAM_MAX: u32 = 8;
-pub const _POSIX_SYMLINK_MAX: u32 = 255;
-pub const _POSIX_SYMLOOP_MAX: u32 = 8;
-pub const _POSIX_TIMER_MAX: u32 = 32;
-pub const _POSIX_TTY_NAME_MAX: u32 = 9;
-pub const _POSIX_TZNAME_MAX: u32 = 6;
-pub const _POSIX_CLOCKRES_MIN: u32 = 20000000;
-pub const NR_OPEN: u32 = 1024;
-pub const NGROUPS_MAX: u32 = 65536;
-pub const ARG_MAX: u32 = 131072;
-pub const LINK_MAX: u32 = 127;
-pub const MAX_CANON: u32 = 255;
-pub const MAX_INPUT: u32 = 255;
-pub const NAME_MAX: u32 = 255;
-pub const PATH_MAX: u32 = 4096;
-pub const PIPE_BUF: u32 = 4096;
-pub const XATTR_NAME_MAX: u32 = 255;
-pub const XATTR_SIZE_MAX: u32 = 65536;
-pub const XATTR_LIST_MAX: u32 = 65536;
-pub const RTSIG_MAX: u32 = 32;
-pub const _POSIX_THREAD_KEYS_MAX: u32 = 128;
-pub const PTHREAD_KEYS_MAX: u32 = 1024;
-pub const _POSIX_THREAD_DESTRUCTOR_ITERATIONS: u32 = 4;
-pub const PTHREAD_DESTRUCTOR_ITERATIONS: u32 = 4;
-pub const _POSIX_THREAD_THREADS_MAX: u32 = 64;
-pub const AIO_PRIO_DELTA_MAX: u32 = 20;
-pub const PTHREAD_STACK_MIN: u32 = 16384;
-pub const DELAYTIMER_MAX: u32 = 2147483647;
-pub const TTY_NAME_MAX: u32 = 32;
-pub const LOGIN_NAME_MAX: u32 = 256;
-pub const HOST_NAME_MAX: u32 = 64;
-pub const MQ_PRIO_MAX: u32 = 32768;
-pub const SEM_VALUE_MAX: u32 = 2147483647;
-pub const _BITS_POSIX2_LIM_H: u32 = 1;
-pub const _POSIX2_BC_BASE_MAX: u32 = 99;
-pub const _POSIX2_BC_DIM_MAX: u32 = 2048;
-pub const _POSIX2_BC_SCALE_MAX: u32 = 99;
-pub const _POSIX2_BC_STRING_MAX: u32 = 1000;
-pub const _POSIX2_COLL_WEIGHTS_MAX: u32 = 2;
-pub const _POSIX2_EXPR_NEST_MAX: u32 = 32;
-pub const _POSIX2_LINE_MAX: u32 = 2048;
-pub const _POSIX2_RE_DUP_MAX: u32 = 255;
-pub const _POSIX2_CHARCLASS_NAME_MAX: u32 = 14;
-pub const BC_BASE_MAX: u32 = 99;
-pub const BC_DIM_MAX: u32 = 2048;
-pub const BC_SCALE_MAX: u32 = 99;
-pub const BC_STRING_MAX: u32 = 1000;
-pub const COLL_WEIGHTS_MAX: u32 = 255;
-pub const EXPR_NEST_MAX: u32 = 32;
-pub const LINE_MAX: u32 = 2048;
-pub const CHARCLASS_NAME_MAX: u32 = 2048;
-pub const RE_DUP_MAX: u32 = 32767;
-pub const _STRING_H: u32 = 1;
-pub const _BITS_TYPES_LOCALE_T_H: u32 = 1;
-pub const _BITS_TYPES___LOCALE_T_H: u32 = 1;
-pub const _STRINGS_H: u32 = 1;
-pub const TF_TSTRING_LITTLE_ENDIAN: u32 = 1;
-pub type wchar_t = ::std::os::raw::c_int;
-pub type __u_char = ::std::os::raw::c_uchar;
-pub type __u_short = ::std::os::raw::c_ushort;
-pub type __u_int = ::std::os::raw::c_uint;
-pub type __u_long = ::std::os::raw::c_ulong;
-pub type __int8_t = ::std::os::raw::c_schar;
-pub type __uint8_t = ::std::os::raw::c_uchar;
-pub type __int16_t = ::std::os::raw::c_short;
-pub type __uint16_t = ::std::os::raw::c_ushort;
-pub type __int32_t = ::std::os::raw::c_int;
-pub type __uint32_t = ::std::os::raw::c_uint;
-pub type __int64_t = ::std::os::raw::c_long;
-pub type __uint64_t = ::std::os::raw::c_ulong;
-pub type __int_least8_t = __int8_t;
-pub type __uint_least8_t = __uint8_t;
-pub type __int_least16_t = __int16_t;
-pub type __uint_least16_t = __uint16_t;
-pub type __int_least32_t = __int32_t;
-pub type __uint_least32_t = __uint32_t;
-pub type __int_least64_t = __int64_t;
-pub type __uint_least64_t = __uint64_t;
-pub type __quad_t = ::std::os::raw::c_long;
-pub type __u_quad_t = ::std::os::raw::c_ulong;
-pub type __intmax_t = ::std::os::raw::c_long;
-pub type __uintmax_t = ::std::os::raw::c_ulong;
-pub type __dev_t = ::std::os::raw::c_ulong;
-pub type __uid_t = ::std::os::raw::c_uint;
-pub type __gid_t = ::std::os::raw::c_uint;
-pub type __ino_t = ::std::os::raw::c_ulong;
-pub type __ino64_t = ::std::os::raw::c_ulong;
-pub type __mode_t = ::std::os::raw::c_uint;
-pub type __nlink_t = ::std::os::raw::c_ulong;
-pub type __off_t = ::std::os::raw::c_long;
-pub type __off64_t = ::std::os::raw::c_long;
-pub type __pid_t = ::std::os::raw::c_int;
-#[repr(C)]
-#[derive(Debug, Copy, Clone)]
-pub struct __fsid_t {
-    pub __val: [::std::os::raw::c_int; 2usize],
-}
-#[test]
-fn bindgen_test_layout___fsid_t() {
-    assert_eq!(
-        ::std::mem::size_of::<__fsid_t>(),
-        8usize,
-        concat!("Size of: ", stringify!(__fsid_t))
-    );
-    assert_eq!(
-        ::std::mem::align_of::<__fsid_t>(),
-        4usize,
-        concat!("Alignment of ", stringify!(__fsid_t))
-    );
-    assert_eq!(
-        unsafe { &(*(::std::ptr::null::<__fsid_t>())).__val as *const _ as usize },
-        0usize,
-        concat!(
-            "Offset of field: ",
-            stringify!(__fsid_t),
-            "::",
-            stringify!(__val)
-        )
-    );
-}
-pub type __clock_t = ::std::os::raw::c_long;
-pub type __rlim_t = ::std::os::raw::c_ulong;
-pub type __rlim64_t = ::std::os::raw::c_ulong;
-pub type __id_t = ::std::os::raw::c_uint;
-pub type __time_t = ::std::os::raw::c_long;
-pub type __useconds_t = ::std::os::raw::c_uint;
-pub type __suseconds_t = ::std::os::raw::c_long;
-pub type __daddr_t = ::std::os::raw::c_int;
-pub type __key_t = ::std::os::raw::c_int;
-pub type __clockid_t = ::std::os::raw::c_int;
-pub type __timer_t = *mut ::std::os::raw::c_void;
-pub type __blksize_t = ::std::os::raw::c_long;
-pub type __blkcnt_t = ::std::os::raw::c_long;
-pub type __blkcnt64_t = ::std::os::raw::c_long;
-pub type __fsblkcnt_t = ::std::os::raw::c_ulong;
-pub type __fsblkcnt64_t = ::std::os::raw::c_ulong;
-pub type __fsfilcnt_t = ::std::os::raw::c_ulong;
-pub type __fsfilcnt64_t = ::std::os::raw::c_ulong;
-pub type __fsword_t = ::std::os::raw::c_long;
-pub type __ssize_t = ::std::os::raw::c_long;
-pub type __syscall_slong_t = ::std::os::raw::c_long;
-pub type __syscall_ulong_t = ::std::os::raw::c_ulong;
-pub type __loff_t = __off64_t;
-pub type __caddr_t = *mut ::std::os::raw::c_char;
-pub type __intptr_t = ::std::os::raw::c_long;
-pub type __socklen_t = ::std::os::raw::c_uint;
-pub type __sig_atomic_t = ::std::os::raw::c_int;
-pub type int_least8_t = __int_least8_t;
-pub type int_least16_t = __int_least16_t;
-pub type int_least32_t = __int_least32_t;
-pub type int_least64_t = __int_least64_t;
-pub type uint_least8_t = __uint_least8_t;
-pub type uint_least16_t = __uint_least16_t;
-pub type uint_least32_t = __uint_least32_t;
-pub type uint_least64_t = __uint_least64_t;
-pub type int_fast8_t = ::std::os::raw::c_schar;
-pub type int_fast16_t = ::std::os::raw::c_long;
-pub type int_fast32_t = ::std::os::raw::c_long;
-pub type int_fast64_t = ::std::os::raw::c_long;
-pub type uint_fast8_t = ::std::os::raw::c_uchar;
-pub type uint_fast16_t = ::std::os::raw::c_ulong;
-pub type uint_fast32_t = ::std::os::raw::c_ulong;
-pub type uint_fast64_t = ::std::os::raw::c_ulong;
-pub type intmax_t = __intmax_t;
-pub type uintmax_t = __uintmax_t;
-#[repr(u32)]
-#[derive(Debug, Copy, Clone, PartialEq, Eq, Hash)]
-pub enum TF_AttrType {
-    TF_ATTR_STRING = 0,
-    TF_ATTR_INT = 1,
-    TF_ATTR_FLOAT = 2,
-    TF_ATTR_BOOL = 3,
-    TF_ATTR_TYPE = 4,
-    TF_ATTR_SHAPE = 5,
-    TF_ATTR_TENSOR = 6,
-    TF_ATTR_PLACEHOLDER = 7,
-    TF_ATTR_FUNC = 8,
-}
-impl TF_DataType {
-    pub const TF_COMPLEX: TF_DataType = TF_DataType::TF_COMPLEX64;
-}
-#[repr(u32)]
-#[derive(Debug, Copy, Clone, PartialEq, Eq, Hash)]
-pub enum TF_DataType {
-    TF_FLOAT = 1,
-    TF_DOUBLE = 2,
-    TF_INT32 = 3,
-    TF_UINT8 = 4,
-    TF_INT16 = 5,
-    TF_INT8 = 6,
-    TF_STRING = 7,
-    TF_COMPLEX64 = 8,
-    TF_INT64 = 9,
-    TF_BOOL = 10,
-    TF_QINT8 = 11,
-    TF_QUINT8 = 12,
-    TF_QINT32 = 13,
-    TF_BFLOAT16 = 14,
-    TF_QINT16 = 15,
-    TF_QUINT16 = 16,
-    TF_UINT16 = 17,
-    TF_COMPLEX128 = 18,
-    TF_HALF = 19,
-    TF_RESOURCE = 20,
-    TF_VARIANT = 21,
-    TF_UINT32 = 22,
-    TF_UINT64 = 23,
-}
-extern "C" {
-    pub fn TF_DataTypeSize(dt: TF_DataType) -> usize;
-}
-#[repr(C)]
-#[derive(Debug, Copy, Clone)]
-pub struct TF_Status {
-    _unused: [u8; 0],
-}
-#[repr(u32)]
-#[derive(Debug, Copy, Clone, PartialEq, Eq, Hash)]
-pub enum TF_Code {
-    TF_OK = 0,
-    TF_CANCELLED = 1,
-    TF_UNKNOWN = 2,
-    TF_INVALID_ARGUMENT = 3,
-    TF_DEADLINE_EXCEEDED = 4,
-    TF_NOT_FOUND = 5,
-    TF_ALREADY_EXISTS = 6,
-    TF_PERMISSION_DENIED = 7,
-    TF_UNAUTHENTICATED = 16,
-    TF_RESOURCE_EXHAUSTED = 8,
-    TF_FAILED_PRECONDITION = 9,
-    TF_ABORTED = 10,
-    TF_OUT_OF_RANGE = 11,
-    TF_UNIMPLEMENTED = 12,
-    TF_INTERNAL = 13,
-    TF_UNAVAILABLE = 14,
-    TF_DATA_LOSS = 15,
-}
-extern "C" {
-    pub fn TF_NewStatus() -> *mut TF_Status;
-}
-extern "C" {
-    pub fn TF_DeleteStatus(arg1: *mut TF_Status);
-}
-extern "C" {
-    pub fn TF_SetStatus(s: *mut TF_Status, code: TF_Code, msg: *const ::std::os::raw::c_char);
-}
-extern "C" {
-    pub fn TF_SetStatusFromIOError(
-        s: *mut TF_Status,
-        error_code: ::std::os::raw::c_int,
-        context: *const ::std::os::raw::c_char,
-    );
-}
-extern "C" {
-    pub fn TF_GetCode(s: *const TF_Status) -> TF_Code;
-}
-extern "C" {
-    pub fn TF_Message(s: *const TF_Status) -> *const ::std::os::raw::c_char;
-}
-#[repr(C)]
-#[derive(Debug, Copy, Clone)]
-pub struct TF_AllocatorAttributes {
-    pub struct_size: usize,
-    pub on_host: ::std::os::raw::c_uchar,
-}
-#[test]
-fn bindgen_test_layout_TF_AllocatorAttributes() {
-    assert_eq!(
-        ::std::mem::size_of::<TF_AllocatorAttributes>(),
-        16usize,
-        concat!("Size of: ", stringify!(TF_AllocatorAttributes))
-    );
-    assert_eq!(
-        ::std::mem::align_of::<TF_AllocatorAttributes>(),
-        8usize,
-        concat!("Alignment of ", stringify!(TF_AllocatorAttributes))
-    );
-    assert_eq!(
-        unsafe {
-            &(*(::std::ptr::null::<TF_AllocatorAttributes>())).struct_size as *const _ as usize
-        },
-        0usize,
-        concat!(
-            "Offset of field: ",
-            stringify!(TF_AllocatorAttributes),
-            "::",
-            stringify!(struct_size)
-        )
-    );
-    assert_eq!(
-        unsafe { &(*(::std::ptr::null::<TF_AllocatorAttributes>())).on_host as *const _ as usize },
-        8usize,
-        concat!(
-            "Offset of field: ",
-            stringify!(TF_AllocatorAttributes),
-            "::",
-            stringify!(on_host)
-        )
-    );
-}
-#[repr(C)]
-#[derive(Debug, Copy, Clone)]
-pub struct TF_Tensor {
-    _unused: [u8; 0],
-}
-extern "C" {
-    pub fn TF_NewTensor(
-        arg1: TF_DataType,
-        dims: *const i64,
-        num_dims: ::std::os::raw::c_int,
-        data: *mut ::std::os::raw::c_void,
-        len: usize,
-        deallocator: ::std::option::Option<
-            unsafe extern "C" fn(
-                data: *mut ::std::os::raw::c_void,
-                len: usize,
-                arg: *mut ::std::os::raw::c_void,
-            ),
-        >,
-        deallocator_arg: *mut ::std::os::raw::c_void,
-    ) -> *mut TF_Tensor;
-}
-extern "C" {
-    pub fn TF_AllocateTensor(
-        arg1: TF_DataType,
-        dims: *const i64,
-        num_dims: ::std::os::raw::c_int,
-        len: usize,
-    ) -> *mut TF_Tensor;
-}
-extern "C" {
-    pub fn TF_TensorMaybeMove(tensor: *mut TF_Tensor) -> *mut TF_Tensor;
-}
-extern "C" {
-    pub fn TF_DeleteTensor(arg1: *mut TF_Tensor);
-}
-extern "C" {
-    pub fn TF_TensorType(arg1: *const TF_Tensor) -> TF_DataType;
-}
-extern "C" {
-    pub fn TF_NumDims(arg1: *const TF_Tensor) -> ::std::os::raw::c_int;
-}
-extern "C" {
-    pub fn TF_Dim(tensor: *const TF_Tensor, dim_index: ::std::os::raw::c_int) -> i64;
-}
-extern "C" {
-    pub fn TF_TensorByteSize(arg1: *const TF_Tensor) -> usize;
-}
-extern "C" {
-    pub fn TF_TensorData(arg1: *const TF_Tensor) -> *mut ::std::os::raw::c_void;
-}
-extern "C" {
-    pub fn TF_TensorElementCount(tensor: *const TF_Tensor) -> i64;
-}
-extern "C" {
-    pub fn TF_TensorBitcastFrom(
-        from: *const TF_Tensor,
-        type_: TF_DataType,
-        to: *mut TF_Tensor,
-        new_dims: *const i64,
-        num_new_dims: ::std::os::raw::c_int,
-        status: *mut TF_Status,
-    );
-}
-extern "C" {
-    pub fn TF_TensorIsAligned(arg1: *const TF_Tensor) -> bool;
-}
-#[repr(u32)]
-#[derive(Debug, Copy, Clone, PartialEq, Eq, Hash)]
-pub enum idtype_t {
-    P_ALL = 0,
-    P_PID = 1,
-    P_PGID = 2,
-}
-pub type _Float32 = f32;
-pub type _Float64 = f64;
-pub type _Float32x = f64;
-pub type _Float64x = u128;
-#[repr(C)]
-#[derive(Debug, Copy, Clone)]
-pub struct div_t {
-    pub quot: ::std::os::raw::c_int,
-    pub rem: ::std::os::raw::c_int,
-}
-#[test]
-fn bindgen_test_layout_div_t() {
-    assert_eq!(
-        ::std::mem::size_of::<div_t>(),
-        8usize,
-        concat!("Size of: ", stringify!(div_t))
-    );
-    assert_eq!(
-        ::std::mem::align_of::<div_t>(),
-        4usize,
-        concat!("Alignment of ", stringify!(div_t))
-    );
-    assert_eq!(
-        unsafe { &(*(::std::ptr::null::<div_t>())).quot as *const _ as usize },
-        0usize,
-        concat!(
-            "Offset of field: ",
-            stringify!(div_t),
-            "::",
-            stringify!(quot)
-        )
-    );
-    assert_eq!(
-        unsafe { &(*(::std::ptr::null::<div_t>())).rem as *const _ as usize },
-        4usize,
-        concat!(
-            "Offset of field: ",
-            stringify!(div_t),
-            "::",
-            stringify!(rem)
-        )
-    );
-}
-#[repr(C)]
-#[derive(Debug, Copy, Clone)]
-pub struct ldiv_t {
-    pub quot: ::std::os::raw::c_long,
-    pub rem: ::std::os::raw::c_long,
-}
-#[test]
-fn bindgen_test_layout_ldiv_t() {
-    assert_eq!(
-        ::std::mem::size_of::<ldiv_t>(),
-        16usize,
-        concat!("Size of: ", stringify!(ldiv_t))
-    );
-    assert_eq!(
-        ::std::mem::align_of::<ldiv_t>(),
-        8usize,
-        concat!("Alignment of ", stringify!(ldiv_t))
-    );
-    assert_eq!(
-        unsafe { &(*(::std::ptr::null::<ldiv_t>())).quot as *const _ as usize },
-        0usize,
-        concat!(
-            "Offset of field: ",
-            stringify!(ldiv_t),
-            "::",
-            stringify!(quot)
-        )
-    );
-    assert_eq!(
-        unsafe { &(*(::std::ptr::null::<ldiv_t>())).rem as *const _ as usize },
-        8usize,
-        concat!(
-            "Offset of field: ",
-            stringify!(ldiv_t),
-            "::",
-            stringify!(rem)
-        )
-    );
-}
-#[repr(C)]
-#[derive(Debug, Copy, Clone)]
-pub struct lldiv_t {
-    pub quot: ::std::os::raw::c_longlong,
-    pub rem: ::std::os::raw::c_longlong,
-}
-#[test]
-fn bindgen_test_layout_lldiv_t() {
-    assert_eq!(
-        ::std::mem::size_of::<lldiv_t>(),
-        16usize,
-        concat!("Size of: ", stringify!(lldiv_t))
-    );
-    assert_eq!(
-        ::std::mem::align_of::<lldiv_t>(),
-        8usize,
-        concat!("Alignment of ", stringify!(lldiv_t))
-    );
-    assert_eq!(
-        unsafe { &(*(::std::ptr::null::<lldiv_t>())).quot as *const _ as usize },
-        0usize,
-        concat!(
-            "Offset of field: ",
-            stringify!(lldiv_t),
-            "::",
-            stringify!(quot)
-        )
-    );
-    assert_eq!(
-        unsafe { &(*(::std::ptr::null::<lldiv_t>())).rem as *const _ as usize },
-        8usize,
-        concat!(
-            "Offset of field: ",
-            stringify!(lldiv_t),
-            "::",
-            stringify!(rem)
-        )
-    );
-}
-extern "C" {
-    pub fn __ctype_get_mb_cur_max() -> usize;
-}
-extern "C" {
-    pub fn atof(__nptr: *const ::std::os::raw::c_char) -> f64;
-}
-extern "C" {
-    pub fn atoi(__nptr: *const ::std::os::raw::c_char) -> ::std::os::raw::c_int;
-}
-extern "C" {
-    pub fn atol(__nptr: *const ::std::os::raw::c_char) -> ::std::os::raw::c_long;
-}
-extern "C" {
-    pub fn atoll(__nptr: *const ::std::os::raw::c_char) -> ::std::os::raw::c_longlong;
-}
-extern "C" {
-    pub fn strtod(
-        __nptr: *const ::std::os::raw::c_char,
-        __endptr: *mut *mut ::std::os::raw::c_char,
-    ) -> f64;
-}
-extern "C" {
-    pub fn strtof(
-        __nptr: *const ::std::os::raw::c_char,
-        __endptr: *mut *mut ::std::os::raw::c_char,
-    ) -> f32;
-}
-extern "C" {
-    pub fn strtold(
-        __nptr: *const ::std::os::raw::c_char,
-        __endptr: *mut *mut ::std::os::raw::c_char,
-    ) -> u128;
-}
-extern "C" {
-    pub fn strtol(
-        __nptr: *const ::std::os::raw::c_char,
-        __endptr: *mut *mut ::std::os::raw::c_char,
-        __base: ::std::os::raw::c_int,
-    ) -> ::std::os::raw::c_long;
-}
-extern "C" {
-    pub fn strtoul(
-        __nptr: *const ::std::os::raw::c_char,
-        __endptr: *mut *mut ::std::os::raw::c_char,
-        __base: ::std::os::raw::c_int,
-    ) -> ::std::os::raw::c_ulong;
-}
-extern "C" {
-    pub fn strtoq(
-        __nptr: *const ::std::os::raw::c_char,
-        __endptr: *mut *mut ::std::os::raw::c_char,
-        __base: ::std::os::raw::c_int,
-    ) -> ::std::os::raw::c_longlong;
-}
-extern "C" {
-    pub fn strtouq(
-        __nptr: *const ::std::os::raw::c_char,
-        __endptr: *mut *mut ::std::os::raw::c_char,
-        __base: ::std::os::raw::c_int,
-    ) -> ::std::os::raw::c_ulonglong;
-}
-extern "C" {
-    pub fn strtoll(
-        __nptr: *const ::std::os::raw::c_char,
-        __endptr: *mut *mut ::std::os::raw::c_char,
-        __base: ::std::os::raw::c_int,
-    ) -> ::std::os::raw::c_longlong;
-}
-extern "C" {
-    pub fn strtoull(
-        __nptr: *const ::std::os::raw::c_char,
-        __endptr: *mut *mut ::std::os::raw::c_char,
-        __base: ::std::os::raw::c_int,
-    ) -> ::std::os::raw::c_ulonglong;
-}
-extern "C" {
-    pub fn l64a(__n: ::std::os::raw::c_long) -> *mut ::std::os::raw::c_char;
-}
-extern "C" {
-    pub fn a64l(__s: *const ::std::os::raw::c_char) -> ::std::os::raw::c_long;
-}
-pub type u_char = __u_char;
-pub type u_short = __u_short;
-pub type u_int = __u_int;
-pub type u_long = __u_long;
-pub type quad_t = __quad_t;
-pub type u_quad_t = __u_quad_t;
-pub type fsid_t = __fsid_t;
-pub type loff_t = __loff_t;
-pub type ino_t = __ino_t;
-pub type dev_t = __dev_t;
-pub type gid_t = __gid_t;
-pub type mode_t = __mode_t;
-pub type nlink_t = __nlink_t;
-pub type uid_t = __uid_t;
-pub type off_t = __off_t;
-pub type pid_t = __pid_t;
-pub type id_t = __id_t;
-pub type daddr_t = __daddr_t;
-pub type caddr_t = __caddr_t;
-pub type key_t = __key_t;
-pub type clock_t = __clock_t;
-pub type clockid_t = __clockid_t;
-pub type time_t = __time_t;
-pub type timer_t = __timer_t;
-pub type ulong = ::std::os::raw::c_ulong;
-pub type ushort = ::std::os::raw::c_ushort;
-pub type uint = ::std::os::raw::c_uint;
-pub type u_int8_t = __uint8_t;
-pub type u_int16_t = __uint16_t;
-pub type u_int32_t = __uint32_t;
-pub type u_int64_t = __uint64_t;
-pub type register_t = ::std::os::raw::c_long;
-#[repr(C)]
-#[derive(Debug, Copy, Clone)]
-pub struct __sigset_t {
-    pub __val: [::std::os::raw::c_ulong; 16usize],
-}
-#[test]
-fn bindgen_test_layout___sigset_t() {
-    assert_eq!(
-        ::std::mem::size_of::<__sigset_t>(),
-        128usize,
-        concat!("Size of: ", stringify!(__sigset_t))
-    );
-    assert_eq!(
-        ::std::mem::align_of::<__sigset_t>(),
-        8usize,
-        concat!("Alignment of ", stringify!(__sigset_t))
-    );
-    assert_eq!(
-        unsafe { &(*(::std::ptr::null::<__sigset_t>())).__val as *const _ as usize },
-        0usize,
-        concat!(
-            "Offset of field: ",
-            stringify!(__sigset_t),
-            "::",
-            stringify!(__val)
-        )
-    );
-}
-pub type sigset_t = __sigset_t;
-#[repr(C)]
-#[derive(Debug, Copy, Clone)]
-pub struct timeval {
-    pub tv_sec: __time_t,
-    pub tv_usec: __suseconds_t,
-}
-#[test]
-fn bindgen_test_layout_timeval() {
-    assert_eq!(
-        ::std::mem::size_of::<timeval>(),
-        16usize,
-        concat!("Size of: ", stringify!(timeval))
-    );
-    assert_eq!(
-        ::std::mem::align_of::<timeval>(),
-        8usize,
-        concat!("Alignment of ", stringify!(timeval))
-    );
-    assert_eq!(
-        unsafe { &(*(::std::ptr::null::<timeval>())).tv_sec as *const _ as usize },
-        0usize,
-        concat!(
-            "Offset of field: ",
-            stringify!(timeval),
-            "::",
-            stringify!(tv_sec)
-        )
-    );
-    assert_eq!(
-        unsafe { &(*(::std::ptr::null::<timeval>())).tv_usec as *const _ as usize },
-        8usize,
-        concat!(
-            "Offset of field: ",
-            stringify!(timeval),
-            "::",
-            stringify!(tv_usec)
-        )
-    );
-}
-#[repr(C)]
-#[derive(Debug, Copy, Clone)]
-pub struct timespec {
-    pub tv_sec: __time_t,
-    pub tv_nsec: __syscall_slong_t,
-}
-#[test]
-fn bindgen_test_layout_timespec() {
-    assert_eq!(
-        ::std::mem::size_of::<timespec>(),
-        16usize,
-        concat!("Size of: ", stringify!(timespec))
-    );
-    assert_eq!(
-        ::std::mem::align_of::<timespec>(),
-        8usize,
-        concat!("Alignment of ", stringify!(timespec))
-    );
-    assert_eq!(
-        unsafe { &(*(::std::ptr::null::<timespec>())).tv_sec as *const _ as usize },
-        0usize,
-        concat!(
-            "Offset of field: ",
-            stringify!(timespec),
-            "::",
-            stringify!(tv_sec)
-        )
-    );
-    assert_eq!(
-        unsafe { &(*(::std::ptr::null::<timespec>())).tv_nsec as *const _ as usize },
-        8usize,
-        concat!(
-            "Offset of field: ",
-            stringify!(timespec),
-            "::",
-            stringify!(tv_nsec)
-        )
-    );
-}
-pub type suseconds_t = __suseconds_t;
-pub type __fd_mask = ::std::os::raw::c_long;
-#[repr(C)]
-#[derive(Debug, Copy, Clone)]
-pub struct fd_set {
-    pub __fds_bits: [__fd_mask; 16usize],
-}
-#[test]
-fn bindgen_test_layout_fd_set() {
-    assert_eq!(
-        ::std::mem::size_of::<fd_set>(),
-        128usize,
-        concat!("Size of: ", stringify!(fd_set))
-    );
-    assert_eq!(
-        ::std::mem::align_of::<fd_set>(),
-        8usize,
-        concat!("Alignment of ", stringify!(fd_set))
-    );
-    assert_eq!(
-        unsafe { &(*(::std::ptr::null::<fd_set>())).__fds_bits as *const _ as usize },
-        0usize,
-        concat!(
-            "Offset of field: ",
-            stringify!(fd_set),
-            "::",
-            stringify!(__fds_bits)
-        )
-    );
-}
-pub type fd_mask = __fd_mask;
-extern "C" {
-    pub fn select(
-        __nfds: ::std::os::raw::c_int,
-        __readfds: *mut fd_set,
-        __writefds: *mut fd_set,
-        __exceptfds: *mut fd_set,
-        __timeout: *mut timeval,
-    ) -> ::std::os::raw::c_int;
-}
-extern "C" {
-    pub fn pselect(
-        __nfds: ::std::os::raw::c_int,
-        __readfds: *mut fd_set,
-        __writefds: *mut fd_set,
-        __exceptfds: *mut fd_set,
-        __timeout: *const timespec,
-        __sigmask: *const __sigset_t,
-    ) -> ::std::os::raw::c_int;
-}
-pub type blksize_t = __blksize_t;
-pub type blkcnt_t = __blkcnt_t;
-pub type fsblkcnt_t = __fsblkcnt_t;
-pub type fsfilcnt_t = __fsfilcnt_t;
-#[repr(C)]
-#[derive(Debug, Copy, Clone)]
-pub struct __pthread_internal_list {
-    pub __prev: *mut __pthread_internal_list,
-    pub __next: *mut __pthread_internal_list,
-}
-#[test]
-fn bindgen_test_layout___pthread_internal_list() {
-    assert_eq!(
-        ::std::mem::size_of::<__pthread_internal_list>(),
-        16usize,
-        concat!("Size of: ", stringify!(__pthread_internal_list))
-    );
-    assert_eq!(
-        ::std::mem::align_of::<__pthread_internal_list>(),
-        8usize,
-        concat!("Alignment of ", stringify!(__pthread_internal_list))
-    );
-    assert_eq!(
-        unsafe { &(*(::std::ptr::null::<__pthread_internal_list>())).__prev as *const _ as usize },
-        0usize,
-        concat!(
-            "Offset of field: ",
-            stringify!(__pthread_internal_list),
-            "::",
-            stringify!(__prev)
-        )
-    );
-    assert_eq!(
-        unsafe { &(*(::std::ptr::null::<__pthread_internal_list>())).__next as *const _ as usize },
-        8usize,
-        concat!(
-            "Offset of field: ",
-            stringify!(__pthread_internal_list),
-            "::",
-            stringify!(__next)
-        )
-    );
-}
-pub type __pthread_list_t = __pthread_internal_list;
-#[repr(C)]
-#[derive(Debug, Copy, Clone)]
-pub struct __pthread_internal_slist {
-    pub __next: *mut __pthread_internal_slist,
-}
-#[test]
-fn bindgen_test_layout___pthread_internal_slist() {
-    assert_eq!(
-        ::std::mem::size_of::<__pthread_internal_slist>(),
-        8usize,
-        concat!("Size of: ", stringify!(__pthread_internal_slist))
-    );
-    assert_eq!(
-        ::std::mem::align_of::<__pthread_internal_slist>(),
-        8usize,
-        concat!("Alignment of ", stringify!(__pthread_internal_slist))
-    );
-    assert_eq!(
-        unsafe { &(*(::std::ptr::null::<__pthread_internal_slist>())).__next as *const _ as usize },
-        0usize,
-        concat!(
-            "Offset of field: ",
-            stringify!(__pthread_internal_slist),
-            "::",
-            stringify!(__next)
-        )
-    );
-}
-pub type __pthread_slist_t = __pthread_internal_slist;
-#[repr(C)]
-#[derive(Debug, Copy, Clone)]
-pub struct __pthread_mutex_s {
-    pub __lock: ::std::os::raw::c_int,
-    pub __count: ::std::os::raw::c_uint,
-    pub __owner: ::std::os::raw::c_int,
-    pub __nusers: ::std::os::raw::c_uint,
-    pub __kind: ::std::os::raw::c_int,
-    pub __spins: ::std::os::raw::c_short,
-    pub __elision: ::std::os::raw::c_short,
-    pub __list: __pthread_list_t,
-}
-#[test]
-fn bindgen_test_layout___pthread_mutex_s() {
-    assert_eq!(
-        ::std::mem::size_of::<__pthread_mutex_s>(),
-        40usize,
-        concat!("Size of: ", stringify!(__pthread_mutex_s))
-    );
-    assert_eq!(
-        ::std::mem::align_of::<__pthread_mutex_s>(),
-        8usize,
-        concat!("Alignment of ", stringify!(__pthread_mutex_s))
-    );
-    assert_eq!(
-        unsafe { &(*(::std::ptr::null::<__pthread_mutex_s>())).__lock as *const _ as usize },
-        0usize,
-        concat!(
-            "Offset of field: ",
-            stringify!(__pthread_mutex_s),
-            "::",
-            stringify!(__lock)
-        )
-    );
-    assert_eq!(
-        unsafe { &(*(::std::ptr::null::<__pthread_mutex_s>())).__count as *const _ as usize },
-        4usize,
-        concat!(
-            "Offset of field: ",
-            stringify!(__pthread_mutex_s),
-            "::",
-            stringify!(__count)
-        )
-    );
-    assert_eq!(
-        unsafe { &(*(::std::ptr::null::<__pthread_mutex_s>())).__owner as *const _ as usize },
-        8usize,
-        concat!(
-            "Offset of field: ",
-            stringify!(__pthread_mutex_s),
-            "::",
-            stringify!(__owner)
-        )
-    );
-    assert_eq!(
-        unsafe { &(*(::std::ptr::null::<__pthread_mutex_s>())).__nusers as *const _ as usize },
-        12usize,
-        concat!(
-            "Offset of field: ",
-            stringify!(__pthread_mutex_s),
-            "::",
-            stringify!(__nusers)
-        )
-    );
-    assert_eq!(
-        unsafe { &(*(::std::ptr::null::<__pthread_mutex_s>())).__kind as *const _ as usize },
-        16usize,
-        concat!(
-            "Offset of field: ",
-            stringify!(__pthread_mutex_s),
-            "::",
-            stringify!(__kind)
-        )
-    );
-    assert_eq!(
-        unsafe { &(*(::std::ptr::null::<__pthread_mutex_s>())).__spins as *const _ as usize },
-        20usize,
-        concat!(
-            "Offset of field: ",
-            stringify!(__pthread_mutex_s),
-            "::",
-            stringify!(__spins)
-        )
-    );
-    assert_eq!(
-        unsafe { &(*(::std::ptr::null::<__pthread_mutex_s>())).__elision as *const _ as usize },
-        22usize,
-        concat!(
-            "Offset of field: ",
-            stringify!(__pthread_mutex_s),
-            "::",
-            stringify!(__elision)
-        )
-    );
-    assert_eq!(
-        unsafe { &(*(::std::ptr::null::<__pthread_mutex_s>())).__list as *const _ as usize },
-        24usize,
-        concat!(
-            "Offset of field: ",
-            stringify!(__pthread_mutex_s),
-            "::",
-            stringify!(__list)
-        )
-    );
-}
-#[repr(C)]
-#[derive(Debug, Copy, Clone)]
-pub struct __pthread_rwlock_arch_t {
-    pub __readers: ::std::os::raw::c_uint,
-    pub __writers: ::std::os::raw::c_uint,
-    pub __wrphase_futex: ::std::os::raw::c_uint,
-    pub __writers_futex: ::std::os::raw::c_uint,
-    pub __pad3: ::std::os::raw::c_uint,
-    pub __pad4: ::std::os::raw::c_uint,
-    pub __cur_writer: ::std::os::raw::c_int,
-    pub __shared: ::std::os::raw::c_int,
-    pub __rwelision: ::std::os::raw::c_schar,
-    pub __pad1: [::std::os::raw::c_uchar; 7usize],
-    pub __pad2: ::std::os::raw::c_ulong,
-    pub __flags: ::std::os::raw::c_uint,
-}
-#[test]
-fn bindgen_test_layout___pthread_rwlock_arch_t() {
-    assert_eq!(
-        ::std::mem::size_of::<__pthread_rwlock_arch_t>(),
-        56usize,
-        concat!("Size of: ", stringify!(__pthread_rwlock_arch_t))
-    );
-    assert_eq!(
-        ::std::mem::align_of::<__pthread_rwlock_arch_t>(),
-        8usize,
-        concat!("Alignment of ", stringify!(__pthread_rwlock_arch_t))
-    );
-    assert_eq!(
-        unsafe {
-            &(*(::std::ptr::null::<__pthread_rwlock_arch_t>())).__readers as *const _ as usize
-        },
-        0usize,
-        concat!(
-            "Offset of field: ",
-            stringify!(__pthread_rwlock_arch_t),
-            "::",
-            stringify!(__readers)
-        )
-    );
-    assert_eq!(
-        unsafe {
-            &(*(::std::ptr::null::<__pthread_rwlock_arch_t>())).__writers as *const _ as usize
-        },
-        4usize,
-        concat!(
-            "Offset of field: ",
-            stringify!(__pthread_rwlock_arch_t),
-            "::",
-            stringify!(__writers)
-        )
-    );
-    assert_eq!(
-        unsafe {
-            &(*(::std::ptr::null::<__pthread_rwlock_arch_t>())).__wrphase_futex as *const _ as usize
-        },
-        8usize,
-        concat!(
-            "Offset of field: ",
-            stringify!(__pthread_rwlock_arch_t),
-            "::",
-            stringify!(__wrphase_futex)
-        )
-    );
-    assert_eq!(
-        unsafe {
-            &(*(::std::ptr::null::<__pthread_rwlock_arch_t>())).__writers_futex as *const _ as usize
-        },
-        12usize,
-        concat!(
-            "Offset of field: ",
-            stringify!(__pthread_rwlock_arch_t),
-            "::",
-            stringify!(__writers_futex)
-        )
-    );
-    assert_eq!(
-        unsafe { &(*(::std::ptr::null::<__pthread_rwlock_arch_t>())).__pad3 as *const _ as usize },
-        16usize,
-        concat!(
-            "Offset of field: ",
-            stringify!(__pthread_rwlock_arch_t),
-            "::",
-            stringify!(__pad3)
-        )
-    );
-    assert_eq!(
-        unsafe { &(*(::std::ptr::null::<__pthread_rwlock_arch_t>())).__pad4 as *const _ as usize },
-        20usize,
-        concat!(
-            "Offset of field: ",
-            stringify!(__pthread_rwlock_arch_t),
-            "::",
-            stringify!(__pad4)
-        )
-    );
-    assert_eq!(
-        unsafe {
-            &(*(::std::ptr::null::<__pthread_rwlock_arch_t>())).__cur_writer as *const _ as usize
-        },
-        24usize,
-        concat!(
-            "Offset of field: ",
-            stringify!(__pthread_rwlock_arch_t),
-            "::",
-            stringify!(__cur_writer)
-        )
-    );
-    assert_eq!(
-        unsafe {
-            &(*(::std::ptr::null::<__pthread_rwlock_arch_t>())).__shared as *const _ as usize
-        },
-        28usize,
-        concat!(
-            "Offset of field: ",
-            stringify!(__pthread_rwlock_arch_t),
-            "::",
-            stringify!(__shared)
-        )
-    );
-    assert_eq!(
-        unsafe {
-            &(*(::std::ptr::null::<__pthread_rwlock_arch_t>())).__rwelision as *const _ as usize
-        },
-        32usize,
-        concat!(
-            "Offset of field: ",
-            stringify!(__pthread_rwlock_arch_t),
-            "::",
-            stringify!(__rwelision)
-        )
-    );
-    assert_eq!(
-        unsafe { &(*(::std::ptr::null::<__pthread_rwlock_arch_t>())).__pad1 as *const _ as usize },
-        33usize,
-        concat!(
-            "Offset of field: ",
-            stringify!(__pthread_rwlock_arch_t),
-            "::",
-            stringify!(__pad1)
-        )
-    );
-    assert_eq!(
-        unsafe { &(*(::std::ptr::null::<__pthread_rwlock_arch_t>())).__pad2 as *const _ as usize },
-        40usize,
-        concat!(
-            "Offset of field: ",
-            stringify!(__pthread_rwlock_arch_t),
-            "::",
-            stringify!(__pad2)
-        )
-    );
-    assert_eq!(
-        unsafe { &(*(::std::ptr::null::<__pthread_rwlock_arch_t>())).__flags as *const _ as usize },
-        48usize,
-        concat!(
-            "Offset of field: ",
-            stringify!(__pthread_rwlock_arch_t),
-            "::",
-            stringify!(__flags)
-        )
-    );
-}
-#[repr(C)]
-#[derive(Copy, Clone)]
-pub struct __pthread_cond_s {
-    pub __bindgen_anon_1: __pthread_cond_s__bindgen_ty_1,
-    pub __bindgen_anon_2: __pthread_cond_s__bindgen_ty_2,
-    pub __g_refs: [::std::os::raw::c_uint; 2usize],
-    pub __g_size: [::std::os::raw::c_uint; 2usize],
-    pub __g1_orig_size: ::std::os::raw::c_uint,
-    pub __wrefs: ::std::os::raw::c_uint,
-    pub __g_signals: [::std::os::raw::c_uint; 2usize],
-}
-#[repr(C)]
-#[derive(Copy, Clone)]
-pub union __pthread_cond_s__bindgen_ty_1 {
-    pub __wseq: ::std::os::raw::c_ulonglong,
-    pub __wseq32: __pthread_cond_s__bindgen_ty_1__bindgen_ty_1,
-    _bindgen_union_align: u64,
-}
-#[repr(C)]
-#[derive(Debug, Copy, Clone)]
-pub struct __pthread_cond_s__bindgen_ty_1__bindgen_ty_1 {
-    pub __low: ::std::os::raw::c_uint,
-    pub __high: ::std::os::raw::c_uint,
-}
-#[test]
-fn bindgen_test_layout___pthread_cond_s__bindgen_ty_1__bindgen_ty_1() {
-    assert_eq!(
-        ::std::mem::size_of::<__pthread_cond_s__bindgen_ty_1__bindgen_ty_1>(),
-        8usize,
-        concat!(
-            "Size of: ",
-            stringify!(__pthread_cond_s__bindgen_ty_1__bindgen_ty_1)
-        )
-    );
-    assert_eq!(
-        ::std::mem::align_of::<__pthread_cond_s__bindgen_ty_1__bindgen_ty_1>(),
-        4usize,
-        concat!(
-            "Alignment of ",
-            stringify!(__pthread_cond_s__bindgen_ty_1__bindgen_ty_1)
-        )
-    );
-    assert_eq!(
-        unsafe {
-            &(*(::std::ptr::null::<__pthread_cond_s__bindgen_ty_1__bindgen_ty_1>())).__low
-                as *const _ as usize
-        },
-        0usize,
-        concat!(
-            "Offset of field: ",
-            stringify!(__pthread_cond_s__bindgen_ty_1__bindgen_ty_1),
-            "::",
-            stringify!(__low)
-        )
-    );
-    assert_eq!(
-        unsafe {
-            &(*(::std::ptr::null::<__pthread_cond_s__bindgen_ty_1__bindgen_ty_1>())).__high
-                as *const _ as usize
-        },
-        4usize,
-        concat!(
-            "Offset of field: ",
-            stringify!(__pthread_cond_s__bindgen_ty_1__bindgen_ty_1),
-            "::",
-            stringify!(__high)
-        )
-    );
-}
-#[test]
-fn bindgen_test_layout___pthread_cond_s__bindgen_ty_1() {
-    assert_eq!(
-        ::std::mem::size_of::<__pthread_cond_s__bindgen_ty_1>(),
-        8usize,
-        concat!("Size of: ", stringify!(__pthread_cond_s__bindgen_ty_1))
-    );
-    assert_eq!(
-        ::std::mem::align_of::<__pthread_cond_s__bindgen_ty_1>(),
-        8usize,
-        concat!("Alignment of ", stringify!(__pthread_cond_s__bindgen_ty_1))
-    );
-    assert_eq!(
-        unsafe {
-            &(*(::std::ptr::null::<__pthread_cond_s__bindgen_ty_1>())).__wseq as *const _ as usize
-        },
-        0usize,
-        concat!(
-            "Offset of field: ",
-            stringify!(__pthread_cond_s__bindgen_ty_1),
-            "::",
-            stringify!(__wseq)
-        )
-    );
-    assert_eq!(
-        unsafe {
-            &(*(::std::ptr::null::<__pthread_cond_s__bindgen_ty_1>())).__wseq32 as *const _ as usize
-        },
-        0usize,
-        concat!(
-            "Offset of field: ",
-            stringify!(__pthread_cond_s__bindgen_ty_1),
-            "::",
-            stringify!(__wseq32)
-        )
-    );
-}
-#[repr(C)]
-#[derive(Copy, Clone)]
-pub union __pthread_cond_s__bindgen_ty_2 {
-    pub __g1_start: ::std::os::raw::c_ulonglong,
-    pub __g1_start32: __pthread_cond_s__bindgen_ty_2__bindgen_ty_1,
-    _bindgen_union_align: u64,
-}
-#[repr(C)]
-#[derive(Debug, Copy, Clone)]
-pub struct __pthread_cond_s__bindgen_ty_2__bindgen_ty_1 {
-    pub __low: ::std::os::raw::c_uint,
-    pub __high: ::std::os::raw::c_uint,
-}
-#[test]
-fn bindgen_test_layout___pthread_cond_s__bindgen_ty_2__bindgen_ty_1() {
-    assert_eq!(
-        ::std::mem::size_of::<__pthread_cond_s__bindgen_ty_2__bindgen_ty_1>(),
-        8usize,
-        concat!(
-            "Size of: ",
-            stringify!(__pthread_cond_s__bindgen_ty_2__bindgen_ty_1)
-        )
-    );
-    assert_eq!(
-        ::std::mem::align_of::<__pthread_cond_s__bindgen_ty_2__bindgen_ty_1>(),
-        4usize,
-        concat!(
-            "Alignment of ",
-            stringify!(__pthread_cond_s__bindgen_ty_2__bindgen_ty_1)
-        )
-    );
-    assert_eq!(
-        unsafe {
-            &(*(::std::ptr::null::<__pthread_cond_s__bindgen_ty_2__bindgen_ty_1>())).__low
-                as *const _ as usize
-        },
-        0usize,
-        concat!(
-            "Offset of field: ",
-            stringify!(__pthread_cond_s__bindgen_ty_2__bindgen_ty_1),
-            "::",
-            stringify!(__low)
-        )
-    );
-    assert_eq!(
-        unsafe {
-            &(*(::std::ptr::null::<__pthread_cond_s__bindgen_ty_2__bindgen_ty_1>())).__high
-                as *const _ as usize
-        },
-        4usize,
-        concat!(
-            "Offset of field: ",
-            stringify!(__pthread_cond_s__bindgen_ty_2__bindgen_ty_1),
-            "::",
-            stringify!(__high)
-        )
-    );
-}
-#[test]
-fn bindgen_test_layout___pthread_cond_s__bindgen_ty_2() {
-    assert_eq!(
-        ::std::mem::size_of::<__pthread_cond_s__bindgen_ty_2>(),
-        8usize,
-        concat!("Size of: ", stringify!(__pthread_cond_s__bindgen_ty_2))
-    );
-    assert_eq!(
-        ::std::mem::align_of::<__pthread_cond_s__bindgen_ty_2>(),
-        8usize,
-        concat!("Alignment of ", stringify!(__pthread_cond_s__bindgen_ty_2))
-    );
-    assert_eq!(
-        unsafe {
-            &(*(::std::ptr::null::<__pthread_cond_s__bindgen_ty_2>())).__g1_start as *const _
-                as usize
-        },
-        0usize,
-        concat!(
-            "Offset of field: ",
-            stringify!(__pthread_cond_s__bindgen_ty_2),
-            "::",
-            stringify!(__g1_start)
-        )
-    );
-    assert_eq!(
-        unsafe {
-            &(*(::std::ptr::null::<__pthread_cond_s__bindgen_ty_2>())).__g1_start32 as *const _
-                as usize
-        },
-        0usize,
-        concat!(
-            "Offset of field: ",
-            stringify!(__pthread_cond_s__bindgen_ty_2),
-            "::",
-            stringify!(__g1_start32)
-        )
-    );
-}
-#[test]
-fn bindgen_test_layout___pthread_cond_s() {
-    assert_eq!(
-        ::std::mem::size_of::<__pthread_cond_s>(),
-        48usize,
-        concat!("Size of: ", stringify!(__pthread_cond_s))
-    );
-    assert_eq!(
-        ::std::mem::align_of::<__pthread_cond_s>(),
-        8usize,
-        concat!("Alignment of ", stringify!(__pthread_cond_s))
-    );
-    assert_eq!(
-        unsafe { &(*(::std::ptr::null::<__pthread_cond_s>())).__g_refs as *const _ as usize },
-        16usize,
-        concat!(
-            "Offset of field: ",
-            stringify!(__pthread_cond_s),
-            "::",
-            stringify!(__g_refs)
-        )
-    );
-    assert_eq!(
-        unsafe { &(*(::std::ptr::null::<__pthread_cond_s>())).__g_size as *const _ as usize },
-        24usize,
-        concat!(
-            "Offset of field: ",
-            stringify!(__pthread_cond_s),
-            "::",
-            stringify!(__g_size)
-        )
-    );
-    assert_eq!(
-        unsafe { &(*(::std::ptr::null::<__pthread_cond_s>())).__g1_orig_size as *const _ as usize },
-        32usize,
-        concat!(
-            "Offset of field: ",
-            stringify!(__pthread_cond_s),
-            "::",
-            stringify!(__g1_orig_size)
-        )
-    );
-    assert_eq!(
-        unsafe { &(*(::std::ptr::null::<__pthread_cond_s>())).__wrefs as *const _ as usize },
-        36usize,
-        concat!(
-            "Offset of field: ",
-            stringify!(__pthread_cond_s),
-            "::",
-            stringify!(__wrefs)
-        )
-    );
-    assert_eq!(
-        unsafe { &(*(::std::ptr::null::<__pthread_cond_s>())).__g_signals as *const _ as usize },
-        40usize,
-        concat!(
-            "Offset of field: ",
-            stringify!(__pthread_cond_s),
-            "::",
-            stringify!(__g_signals)
-        )
-    );
-}
-pub type pthread_t = ::std::os::raw::c_ulong;
-#[repr(C)]
-#[derive(Copy, Clone)]
-pub union pthread_mutexattr_t {
-    pub __size: [::std::os::raw::c_char; 4usize],
-    pub __align: ::std::os::raw::c_int,
-    _bindgen_union_align: u32,
-}
-#[test]
-fn bindgen_test_layout_pthread_mutexattr_t() {
-    assert_eq!(
-        ::std::mem::size_of::<pthread_mutexattr_t>(),
-        4usize,
-        concat!("Size of: ", stringify!(pthread_mutexattr_t))
-    );
-    assert_eq!(
-        ::std::mem::align_of::<pthread_mutexattr_t>(),
-        4usize,
-        concat!("Alignment of ", stringify!(pthread_mutexattr_t))
-    );
-    assert_eq!(
-        unsafe { &(*(::std::ptr::null::<pthread_mutexattr_t>())).__size as *const _ as usize },
-        0usize,
-        concat!(
-            "Offset of field: ",
-            stringify!(pthread_mutexattr_t),
-            "::",
-            stringify!(__size)
-        )
-    );
-    assert_eq!(
-        unsafe { &(*(::std::ptr::null::<pthread_mutexattr_t>())).__align as *const _ as usize },
-        0usize,
-        concat!(
-            "Offset of field: ",
-            stringify!(pthread_mutexattr_t),
-            "::",
-            stringify!(__align)
-        )
-    );
-}
-#[repr(C)]
-#[derive(Copy, Clone)]
-pub union pthread_condattr_t {
-    pub __size: [::std::os::raw::c_char; 4usize],
-    pub __align: ::std::os::raw::c_int,
-    _bindgen_union_align: u32,
-}
-#[test]
-fn bindgen_test_layout_pthread_condattr_t() {
-    assert_eq!(
-        ::std::mem::size_of::<pthread_condattr_t>(),
-        4usize,
-        concat!("Size of: ", stringify!(pthread_condattr_t))
-    );
-    assert_eq!(
-        ::std::mem::align_of::<pthread_condattr_t>(),
-        4usize,
-        concat!("Alignment of ", stringify!(pthread_condattr_t))
-    );
-    assert_eq!(
-        unsafe { &(*(::std::ptr::null::<pthread_condattr_t>())).__size as *const _ as usize },
-        0usize,
-        concat!(
-            "Offset of field: ",
-            stringify!(pthread_condattr_t),
-            "::",
-            stringify!(__size)
-        )
-    );
-    assert_eq!(
-        unsafe { &(*(::std::ptr::null::<pthread_condattr_t>())).__align as *const _ as usize },
-        0usize,
-        concat!(
-            "Offset of field: ",
-            stringify!(pthread_condattr_t),
-            "::",
-            stringify!(__align)
-        )
-    );
-}
-pub type pthread_key_t = ::std::os::raw::c_uint;
-pub type pthread_once_t = ::std::os::raw::c_int;
-#[repr(C)]
-#[derive(Copy, Clone)]
-pub union pthread_attr_t {
-    pub __size: [::std::os::raw::c_char; 56usize],
-    pub __align: ::std::os::raw::c_long,
-    _bindgen_union_align: [u64; 7usize],
-}
-#[test]
-fn bindgen_test_layout_pthread_attr_t() {
-    assert_eq!(
-        ::std::mem::size_of::<pthread_attr_t>(),
-        56usize,
-        concat!("Size of: ", stringify!(pthread_attr_t))
-    );
-    assert_eq!(
-        ::std::mem::align_of::<pthread_attr_t>(),
-        8usize,
-        concat!("Alignment of ", stringify!(pthread_attr_t))
-    );
-    assert_eq!(
-        unsafe { &(*(::std::ptr::null::<pthread_attr_t>())).__size as *const _ as usize },
-        0usize,
-        concat!(
-            "Offset of field: ",
-            stringify!(pthread_attr_t),
-            "::",
-            stringify!(__size)
-        )
-    );
-    assert_eq!(
-        unsafe { &(*(::std::ptr::null::<pthread_attr_t>())).__align as *const _ as usize },
-        0usize,
-        concat!(
-            "Offset of field: ",
-            stringify!(pthread_attr_t),
-            "::",
-            stringify!(__align)
-        )
-    );
-}
-#[repr(C)]
-#[derive(Copy, Clone)]
-pub union pthread_mutex_t {
-    pub __data: __pthread_mutex_s,
-    pub __size: [::std::os::raw::c_char; 40usize],
-    pub __align: ::std::os::raw::c_long,
-    _bindgen_union_align: [u64; 5usize],
-}
-#[test]
-fn bindgen_test_layout_pthread_mutex_t() {
-    assert_eq!(
-        ::std::mem::size_of::<pthread_mutex_t>(),
-        40usize,
-        concat!("Size of: ", stringify!(pthread_mutex_t))
-    );
-    assert_eq!(
-        ::std::mem::align_of::<pthread_mutex_t>(),
-        8usize,
-        concat!("Alignment of ", stringify!(pthread_mutex_t))
-    );
-    assert_eq!(
-        unsafe { &(*(::std::ptr::null::<pthread_mutex_t>())).__data as *const _ as usize },
-        0usize,
-        concat!(
-            "Offset of field: ",
-            stringify!(pthread_mutex_t),
-            "::",
-            stringify!(__data)
-        )
-    );
-    assert_eq!(
-        unsafe { &(*(::std::ptr::null::<pthread_mutex_t>())).__size as *const _ as usize },
-        0usize,
-        concat!(
-            "Offset of field: ",
-            stringify!(pthread_mutex_t),
-            "::",
-            stringify!(__size)
-        )
-    );
-    assert_eq!(
-        unsafe { &(*(::std::ptr::null::<pthread_mutex_t>())).__align as *const _ as usize },
-        0usize,
-        concat!(
-            "Offset of field: ",
-            stringify!(pthread_mutex_t),
-            "::",
-            stringify!(__align)
-        )
-    );
-}
-#[repr(C)]
-#[derive(Copy, Clone)]
-pub union pthread_cond_t {
-    pub __data: __pthread_cond_s,
-    pub __size: [::std::os::raw::c_char; 48usize],
-    pub __align: ::std::os::raw::c_longlong,
-    _bindgen_union_align: [u64; 6usize],
-}
-#[test]
-fn bindgen_test_layout_pthread_cond_t() {
-    assert_eq!(
-        ::std::mem::size_of::<pthread_cond_t>(),
-        48usize,
-        concat!("Size of: ", stringify!(pthread_cond_t))
-    );
-    assert_eq!(
-        ::std::mem::align_of::<pthread_cond_t>(),
-        8usize,
-        concat!("Alignment of ", stringify!(pthread_cond_t))
-    );
-    assert_eq!(
-        unsafe { &(*(::std::ptr::null::<pthread_cond_t>())).__data as *const _ as usize },
-        0usize,
-        concat!(
-            "Offset of field: ",
-            stringify!(pthread_cond_t),
-            "::",
-            stringify!(__data)
-        )
-    );
-    assert_eq!(
-        unsafe { &(*(::std::ptr::null::<pthread_cond_t>())).__size as *const _ as usize },
-        0usize,
-        concat!(
-            "Offset of field: ",
-            stringify!(pthread_cond_t),
-            "::",
-            stringify!(__size)
-        )
-    );
-    assert_eq!(
-        unsafe { &(*(::std::ptr::null::<pthread_cond_t>())).__align as *const _ as usize },
-        0usize,
-        concat!(
-            "Offset of field: ",
-            stringify!(pthread_cond_t),
-            "::",
-            stringify!(__align)
-        )
-    );
-}
-#[repr(C)]
-#[derive(Copy, Clone)]
-pub union pthread_rwlock_t {
-    pub __data: __pthread_rwlock_arch_t,
-    pub __size: [::std::os::raw::c_char; 56usize],
-    pub __align: ::std::os::raw::c_long,
-    _bindgen_union_align: [u64; 7usize],
-}
-#[test]
-fn bindgen_test_layout_pthread_rwlock_t() {
-    assert_eq!(
-        ::std::mem::size_of::<pthread_rwlock_t>(),
-        56usize,
-        concat!("Size of: ", stringify!(pthread_rwlock_t))
-    );
-    assert_eq!(
-        ::std::mem::align_of::<pthread_rwlock_t>(),
-        8usize,
-        concat!("Alignment of ", stringify!(pthread_rwlock_t))
-    );
-    assert_eq!(
-        unsafe { &(*(::std::ptr::null::<pthread_rwlock_t>())).__data as *const _ as usize },
-        0usize,
-        concat!(
-            "Offset of field: ",
-            stringify!(pthread_rwlock_t),
-            "::",
-            stringify!(__data)
-        )
-    );
-    assert_eq!(
-        unsafe { &(*(::std::ptr::null::<pthread_rwlock_t>())).__size as *const _ as usize },
-        0usize,
-        concat!(
-            "Offset of field: ",
-            stringify!(pthread_rwlock_t),
-            "::",
-            stringify!(__size)
-        )
-    );
-    assert_eq!(
-        unsafe { &(*(::std::ptr::null::<pthread_rwlock_t>())).__align as *const _ as usize },
-        0usize,
-        concat!(
-            "Offset of field: ",
-            stringify!(pthread_rwlock_t),
-            "::",
-            stringify!(__align)
-        )
-    );
-}
-#[repr(C)]
-#[derive(Copy, Clone)]
-pub union pthread_rwlockattr_t {
-    pub __size: [::std::os::raw::c_char; 8usize],
-    pub __align: ::std::os::raw::c_long,
-    _bindgen_union_align: u64,
-}
-#[test]
-fn bindgen_test_layout_pthread_rwlockattr_t() {
-    assert_eq!(
-        ::std::mem::size_of::<pthread_rwlockattr_t>(),
-        8usize,
-        concat!("Size of: ", stringify!(pthread_rwlockattr_t))
-    );
-    assert_eq!(
-        ::std::mem::align_of::<pthread_rwlockattr_t>(),
-        8usize,
-        concat!("Alignment of ", stringify!(pthread_rwlockattr_t))
-    );
-    assert_eq!(
-        unsafe { &(*(::std::ptr::null::<pthread_rwlockattr_t>())).__size as *const _ as usize },
-        0usize,
-        concat!(
-            "Offset of field: ",
-            stringify!(pthread_rwlockattr_t),
-            "::",
-            stringify!(__size)
-        )
-    );
-    assert_eq!(
-        unsafe { &(*(::std::ptr::null::<pthread_rwlockattr_t>())).__align as *const _ as usize },
-        0usize,
-        concat!(
-            "Offset of field: ",
-            stringify!(pthread_rwlockattr_t),
-            "::",
-            stringify!(__align)
-        )
-    );
-}
-pub type pthread_spinlock_t = ::std::os::raw::c_int;
-#[repr(C)]
-#[derive(Copy, Clone)]
-pub union pthread_barrier_t {
-    pub __size: [::std::os::raw::c_char; 32usize],
-    pub __align: ::std::os::raw::c_long,
-    _bindgen_union_align: [u64; 4usize],
-}
-#[test]
-fn bindgen_test_layout_pthread_barrier_t() {
-    assert_eq!(
-        ::std::mem::size_of::<pthread_barrier_t>(),
-        32usize,
-        concat!("Size of: ", stringify!(pthread_barrier_t))
-    );
-    assert_eq!(
-        ::std::mem::align_of::<pthread_barrier_t>(),
-        8usize,
-        concat!("Alignment of ", stringify!(pthread_barrier_t))
-    );
-    assert_eq!(
-        unsafe { &(*(::std::ptr::null::<pthread_barrier_t>())).__size as *const _ as usize },
-        0usize,
-        concat!(
-            "Offset of field: ",
-            stringify!(pthread_barrier_t),
-            "::",
-            stringify!(__size)
-        )
-    );
-    assert_eq!(
-        unsafe { &(*(::std::ptr::null::<pthread_barrier_t>())).__align as *const _ as usize },
-        0usize,
-        concat!(
-            "Offset of field: ",
-            stringify!(pthread_barrier_t),
-            "::",
-            stringify!(__align)
-        )
-    );
-}
-#[repr(C)]
-#[derive(Copy, Clone)]
-pub union pthread_barrierattr_t {
-    pub __size: [::std::os::raw::c_char; 4usize],
-    pub __align: ::std::os::raw::c_int,
-    _bindgen_union_align: u32,
-}
-#[test]
-fn bindgen_test_layout_pthread_barrierattr_t() {
-    assert_eq!(
-        ::std::mem::size_of::<pthread_barrierattr_t>(),
-        4usize,
-        concat!("Size of: ", stringify!(pthread_barrierattr_t))
-    );
-    assert_eq!(
-        ::std::mem::align_of::<pthread_barrierattr_t>(),
-        4usize,
-        concat!("Alignment of ", stringify!(pthread_barrierattr_t))
-    );
-    assert_eq!(
-        unsafe { &(*(::std::ptr::null::<pthread_barrierattr_t>())).__size as *const _ as usize },
-        0usize,
-        concat!(
-            "Offset of field: ",
-            stringify!(pthread_barrierattr_t),
-            "::",
-            stringify!(__size)
-        )
-    );
-    assert_eq!(
-        unsafe { &(*(::std::ptr::null::<pthread_barrierattr_t>())).__align as *const _ as usize },
-        0usize,
-        concat!(
-            "Offset of field: ",
-            stringify!(pthread_barrierattr_t),
-            "::",
-            stringify!(__align)
-        )
-    );
-}
-extern "C" {
-    pub fn random() -> ::std::os::raw::c_long;
-}
-extern "C" {
-    pub fn srandom(__seed: ::std::os::raw::c_uint);
-}
-extern "C" {
-    pub fn initstate(
-        __seed: ::std::os::raw::c_uint,
-        __statebuf: *mut ::std::os::raw::c_char,
-        __statelen: usize,
-    ) -> *mut ::std::os::raw::c_char;
-}
-extern "C" {
-    pub fn setstate(__statebuf: *mut ::std::os::raw::c_char) -> *mut ::std::os::raw::c_char;
-}
-#[repr(C)]
-#[derive(Debug, Copy, Clone)]
-pub struct random_data {
-    pub fptr: *mut i32,
-    pub rptr: *mut i32,
-    pub state: *mut i32,
-    pub rand_type: ::std::os::raw::c_int,
-    pub rand_deg: ::std::os::raw::c_int,
-    pub rand_sep: ::std::os::raw::c_int,
-    pub end_ptr: *mut i32,
-}
-#[test]
-fn bindgen_test_layout_random_data() {
-    assert_eq!(
-        ::std::mem::size_of::<random_data>(),
-        48usize,
-        concat!("Size of: ", stringify!(random_data))
-    );
-    assert_eq!(
-        ::std::mem::align_of::<random_data>(),
-        8usize,
-        concat!("Alignment of ", stringify!(random_data))
-    );
-    assert_eq!(
-        unsafe { &(*(::std::ptr::null::<random_data>())).fptr as *const _ as usize },
-        0usize,
-        concat!(
-            "Offset of field: ",
-            stringify!(random_data),
-            "::",
-            stringify!(fptr)
-        )
-    );
-    assert_eq!(
-        unsafe { &(*(::std::ptr::null::<random_data>())).rptr as *const _ as usize },
-        8usize,
-        concat!(
-            "Offset of field: ",
-            stringify!(random_data),
-            "::",
-            stringify!(rptr)
-        )
-    );
-    assert_eq!(
-        unsafe { &(*(::std::ptr::null::<random_data>())).state as *const _ as usize },
-        16usize,
-        concat!(
-            "Offset of field: ",
-            stringify!(random_data),
-            "::",
-            stringify!(state)
-        )
-    );
-    assert_eq!(
-        unsafe { &(*(::std::ptr::null::<random_data>())).rand_type as *const _ as usize },
-        24usize,
-        concat!(
-            "Offset of field: ",
-            stringify!(random_data),
-            "::",
-            stringify!(rand_type)
-        )
-    );
-    assert_eq!(
-        unsafe { &(*(::std::ptr::null::<random_data>())).rand_deg as *const _ as usize },
-        28usize,
-        concat!(
-            "Offset of field: ",
-            stringify!(random_data),
-            "::",
-            stringify!(rand_deg)
-        )
-    );
-    assert_eq!(
-        unsafe { &(*(::std::ptr::null::<random_data>())).rand_sep as *const _ as usize },
-        32usize,
-        concat!(
-            "Offset of field: ",
-            stringify!(random_data),
-            "::",
-            stringify!(rand_sep)
-        )
-    );
-    assert_eq!(
-        unsafe { &(*(::std::ptr::null::<random_data>())).end_ptr as *const _ as usize },
-        40usize,
-        concat!(
-            "Offset of field: ",
-            stringify!(random_data),
-            "::",
-            stringify!(end_ptr)
-        )
-    );
-}
-extern "C" {
-    pub fn random_r(__buf: *mut random_data, __result: *mut i32) -> ::std::os::raw::c_int;
-}
-extern "C" {
-    pub fn srandom_r(
-        __seed: ::std::os::raw::c_uint,
-        __buf: *mut random_data,
-    ) -> ::std::os::raw::c_int;
-}
-extern "C" {
-    pub fn initstate_r(
-        __seed: ::std::os::raw::c_uint,
-        __statebuf: *mut ::std::os::raw::c_char,
-        __statelen: usize,
-        __buf: *mut random_data,
-    ) -> ::std::os::raw::c_int;
-}
-extern "C" {
-    pub fn setstate_r(
-        __statebuf: *mut ::std::os::raw::c_char,
-        __buf: *mut random_data,
-    ) -> ::std::os::raw::c_int;
-}
-extern "C" {
-    pub fn rand() -> ::std::os::raw::c_int;
-}
-extern "C" {
-    pub fn srand(__seed: ::std::os::raw::c_uint);
-}
-extern "C" {
-    pub fn rand_r(__seed: *mut ::std::os::raw::c_uint) -> ::std::os::raw::c_int;
-}
-extern "C" {
-    pub fn drand48() -> f64;
-}
-extern "C" {
-    pub fn erand48(__xsubi: *mut ::std::os::raw::c_ushort) -> f64;
-}
-extern "C" {
-    pub fn lrand48() -> ::std::os::raw::c_long;
-}
-extern "C" {
-    pub fn nrand48(__xsubi: *mut ::std::os::raw::c_ushort) -> ::std::os::raw::c_long;
-}
-extern "C" {
-    pub fn mrand48() -> ::std::os::raw::c_long;
-}
-extern "C" {
-    pub fn jrand48(__xsubi: *mut ::std::os::raw::c_ushort) -> ::std::os::raw::c_long;
-}
-extern "C" {
-    pub fn srand48(__seedval: ::std::os::raw::c_long);
-}
-extern "C" {
-    pub fn seed48(__seed16v: *mut ::std::os::raw::c_ushort) -> *mut ::std::os::raw::c_ushort;
-}
-extern "C" {
-    pub fn lcong48(__param: *mut ::std::os::raw::c_ushort);
-}
-#[repr(C)]
-#[derive(Debug, Copy, Clone)]
-pub struct drand48_data {
-    pub __x: [::std::os::raw::c_ushort; 3usize],
-    pub __old_x: [::std::os::raw::c_ushort; 3usize],
-    pub __c: ::std::os::raw::c_ushort,
-    pub __init: ::std::os::raw::c_ushort,
-    pub __a: ::std::os::raw::c_ulonglong,
-}
-#[test]
-fn bindgen_test_layout_drand48_data() {
-    assert_eq!(
-        ::std::mem::size_of::<drand48_data>(),
-        24usize,
-        concat!("Size of: ", stringify!(drand48_data))
-    );
-    assert_eq!(
-        ::std::mem::align_of::<drand48_data>(),
-        8usize,
-        concat!("Alignment of ", stringify!(drand48_data))
-    );
-    assert_eq!(
-        unsafe { &(*(::std::ptr::null::<drand48_data>())).__x as *const _ as usize },
-        0usize,
-        concat!(
-            "Offset of field: ",
-            stringify!(drand48_data),
-            "::",
-            stringify!(__x)
-        )
-    );
-    assert_eq!(
-        unsafe { &(*(::std::ptr::null::<drand48_data>())).__old_x as *const _ as usize },
-        6usize,
-        concat!(
-            "Offset of field: ",
-            stringify!(drand48_data),
-            "::",
-            stringify!(__old_x)
-        )
-    );
-    assert_eq!(
-        unsafe { &(*(::std::ptr::null::<drand48_data>())).__c as *const _ as usize },
-        12usize,
-        concat!(
-            "Offset of field: ",
-            stringify!(drand48_data),
-            "::",
-            stringify!(__c)
-        )
-    );
-    assert_eq!(
-        unsafe { &(*(::std::ptr::null::<drand48_data>())).__init as *const _ as usize },
-        14usize,
-        concat!(
-            "Offset of field: ",
-            stringify!(drand48_data),
-            "::",
-            stringify!(__init)
-        )
-    );
-    assert_eq!(
-        unsafe { &(*(::std::ptr::null::<drand48_data>())).__a as *const _ as usize },
-        16usize,
-        concat!(
-            "Offset of field: ",
-            stringify!(drand48_data),
-            "::",
-            stringify!(__a)
-        )
-    );
-}
-extern "C" {
-    pub fn drand48_r(__buffer: *mut drand48_data, __result: *mut f64) -> ::std::os::raw::c_int;
-}
-extern "C" {
-    pub fn erand48_r(
-        __xsubi: *mut ::std::os::raw::c_ushort,
-        __buffer: *mut drand48_data,
-        __result: *mut f64,
-    ) -> ::std::os::raw::c_int;
-}
-extern "C" {
-    pub fn lrand48_r(
-        __buffer: *mut drand48_data,
-        __result: *mut ::std::os::raw::c_long,
-    ) -> ::std::os::raw::c_int;
-}
-extern "C" {
-    pub fn nrand48_r(
-        __xsubi: *mut ::std::os::raw::c_ushort,
-        __buffer: *mut drand48_data,
-        __result: *mut ::std::os::raw::c_long,
-    ) -> ::std::os::raw::c_int;
-}
-extern "C" {
-    pub fn mrand48_r(
-        __buffer: *mut drand48_data,
-        __result: *mut ::std::os::raw::c_long,
-    ) -> ::std::os::raw::c_int;
-}
-extern "C" {
-    pub fn jrand48_r(
-        __xsubi: *mut ::std::os::raw::c_ushort,
-        __buffer: *mut drand48_data,
-        __result: *mut ::std::os::raw::c_long,
-    ) -> ::std::os::raw::c_int;
-}
-extern "C" {
-    pub fn srand48_r(
-        __seedval: ::std::os::raw::c_long,
-        __buffer: *mut drand48_data,
-    ) -> ::std::os::raw::c_int;
-}
-extern "C" {
-    pub fn seed48_r(
-        __seed16v: *mut ::std::os::raw::c_ushort,
-        __buffer: *mut drand48_data,
-    ) -> ::std::os::raw::c_int;
-}
-extern "C" {
-    pub fn lcong48_r(
-        __param: *mut ::std::os::raw::c_ushort,
-        __buffer: *mut drand48_data,
-    ) -> ::std::os::raw::c_int;
-}
-extern "C" {
-    pub fn malloc(__size: ::std::os::raw::c_ulong) -> *mut ::std::os::raw::c_void;
-}
-extern "C" {
-    pub fn calloc(
-        __nmemb: ::std::os::raw::c_ulong,
-        __size: ::std::os::raw::c_ulong,
-    ) -> *mut ::std::os::raw::c_void;
-}
-extern "C" {
-    pub fn realloc(
-        __ptr: *mut ::std::os::raw::c_void,
-        __size: ::std::os::raw::c_ulong,
-    ) -> *mut ::std::os::raw::c_void;
-}
-extern "C" {
-    pub fn reallocarray(
-        __ptr: *mut ::std::os::raw::c_void,
-        __nmemb: usize,
-        __size: usize,
-    ) -> *mut ::std::os::raw::c_void;
-}
-extern "C" {
-    pub fn free(__ptr: *mut ::std::os::raw::c_void);
-}
-extern "C" {
-    pub fn alloca(__size: ::std::os::raw::c_ulong) -> *mut ::std::os::raw::c_void;
-}
-extern "C" {
-    pub fn valloc(__size: usize) -> *mut ::std::os::raw::c_void;
-}
-extern "C" {
-    pub fn posix_memalign(
-        __memptr: *mut *mut ::std::os::raw::c_void,
-        __alignment: usize,
-        __size: usize,
-    ) -> ::std::os::raw::c_int;
-}
-extern "C" {
-    pub fn aligned_alloc(__alignment: usize, __size: usize) -> *mut ::std::os::raw::c_void;
-}
-extern "C" {
-    pub fn abort();
-}
-extern "C" {
-    pub fn atexit(__func: ::std::option::Option<unsafe extern "C" fn()>) -> ::std::os::raw::c_int;
-}
-extern "C" {
-    pub fn at_quick_exit(
-        __func: ::std::option::Option<unsafe extern "C" fn()>,
-    ) -> ::std::os::raw::c_int;
-}
-extern "C" {
-    pub fn on_exit(
-        __func: ::std::option::Option<
-            unsafe extern "C" fn(
-                __status: ::std::os::raw::c_int,
-                __arg: *mut ::std::os::raw::c_void,
-            ),
-        >,
-        __arg: *mut ::std::os::raw::c_void,
-    ) -> ::std::os::raw::c_int;
-}
-extern "C" {
-    pub fn exit(__status: ::std::os::raw::c_int);
-}
-extern "C" {
-    pub fn quick_exit(__status: ::std::os::raw::c_int);
-}
-extern "C" {
-    pub fn _Exit(__status: ::std::os::raw::c_int);
-}
-extern "C" {
-    pub fn getenv(__name: *const ::std::os::raw::c_char) -> *mut ::std::os::raw::c_char;
-}
-extern "C" {
-    pub fn putenv(__string: *mut ::std::os::raw::c_char) -> ::std::os::raw::c_int;
-}
-extern "C" {
-    pub fn setenv(
-        __name: *const ::std::os::raw::c_char,
-        __value: *const ::std::os::raw::c_char,
-        __replace: ::std::os::raw::c_int,
-    ) -> ::std::os::raw::c_int;
-}
-extern "C" {
-    pub fn unsetenv(__name: *const ::std::os::raw::c_char) -> ::std::os::raw::c_int;
-}
-extern "C" {
-    pub fn clearenv() -> ::std::os::raw::c_int;
-}
-extern "C" {
-    pub fn mktemp(__template: *mut ::std::os::raw::c_char) -> *mut ::std::os::raw::c_char;
-}
-extern "C" {
-    pub fn mkstemp(__template: *mut ::std::os::raw::c_char) -> ::std::os::raw::c_int;
-}
-extern "C" {
-    pub fn mkstemps(
-        __template: *mut ::std::os::raw::c_char,
-        __suffixlen: ::std::os::raw::c_int,
-    ) -> ::std::os::raw::c_int;
-}
-extern "C" {
-    pub fn mkdtemp(__template: *mut ::std::os::raw::c_char) -> *mut ::std::os::raw::c_char;
-}
-extern "C" {
-    pub fn system(__command: *const ::std::os::raw::c_char) -> ::std::os::raw::c_int;
-}
-extern "C" {
-    pub fn realpath(
-        __name: *const ::std::os::raw::c_char,
-        __resolved: *mut ::std::os::raw::c_char,
-    ) -> *mut ::std::os::raw::c_char;
-}
-pub type __compar_fn_t = ::std::option::Option<
-    unsafe extern "C" fn(
-        arg1: *const ::std::os::raw::c_void,
-        arg2: *const ::std::os::raw::c_void,
-    ) -> ::std::os::raw::c_int,
->;
-extern "C" {
-    pub fn bsearch(
-        __key: *const ::std::os::raw::c_void,
-        __base: *const ::std::os::raw::c_void,
-        __nmemb: usize,
-        __size: usize,
-        __compar: __compar_fn_t,
-    ) -> *mut ::std::os::raw::c_void;
-}
-extern "C" {
-    pub fn qsort(
-        __base: *mut ::std::os::raw::c_void,
-        __nmemb: usize,
-        __size: usize,
-        __compar: __compar_fn_t,
-    );
-}
-extern "C" {
-    pub fn abs(__x: ::std::os::raw::c_int) -> ::std::os::raw::c_int;
-}
-extern "C" {
-    pub fn labs(__x: ::std::os::raw::c_long) -> ::std::os::raw::c_long;
-}
-extern "C" {
-    pub fn llabs(__x: ::std::os::raw::c_longlong) -> ::std::os::raw::c_longlong;
-}
-extern "C" {
-    pub fn div(__numer: ::std::os::raw::c_int, __denom: ::std::os::raw::c_int) -> div_t;
-}
-extern "C" {
-    pub fn ldiv(__numer: ::std::os::raw::c_long, __denom: ::std::os::raw::c_long) -> ldiv_t;
-}
-extern "C" {
-    pub fn lldiv(
-        __numer: ::std::os::raw::c_longlong,
-        __denom: ::std::os::raw::c_longlong,
-    ) -> lldiv_t;
-}
-extern "C" {
-    pub fn ecvt(
-        __value: f64,
-        __ndigit: ::std::os::raw::c_int,
-        __decpt: *mut ::std::os::raw::c_int,
-        __sign: *mut ::std::os::raw::c_int,
-    ) -> *mut ::std::os::raw::c_char;
-}
-extern "C" {
-    pub fn fcvt(
-        __value: f64,
-        __ndigit: ::std::os::raw::c_int,
-        __decpt: *mut ::std::os::raw::c_int,
-        __sign: *mut ::std::os::raw::c_int,
-    ) -> *mut ::std::os::raw::c_char;
-}
-extern "C" {
-    pub fn gcvt(
-        __value: f64,
-        __ndigit: ::std::os::raw::c_int,
-        __buf: *mut ::std::os::raw::c_char,
-    ) -> *mut ::std::os::raw::c_char;
-}
-extern "C" {
-    pub fn qecvt(
-        __value: u128,
-        __ndigit: ::std::os::raw::c_int,
-        __decpt: *mut ::std::os::raw::c_int,
-        __sign: *mut ::std::os::raw::c_int,
-    ) -> *mut ::std::os::raw::c_char;
-}
-extern "C" {
-    pub fn qfcvt(
-        __value: u128,
-        __ndigit: ::std::os::raw::c_int,
-        __decpt: *mut ::std::os::raw::c_int,
-        __sign: *mut ::std::os::raw::c_int,
-    ) -> *mut ::std::os::raw::c_char;
-}
-extern "C" {
-    pub fn qgcvt(
-        __value: u128,
-        __ndigit: ::std::os::raw::c_int,
-        __buf: *mut ::std::os::raw::c_char,
-    ) -> *mut ::std::os::raw::c_char;
-}
-extern "C" {
-    pub fn ecvt_r(
-        __value: f64,
-        __ndigit: ::std::os::raw::c_int,
-        __decpt: *mut ::std::os::raw::c_int,
-        __sign: *mut ::std::os::raw::c_int,
-        __buf: *mut ::std::os::raw::c_char,
-        __len: usize,
-    ) -> ::std::os::raw::c_int;
-}
-extern "C" {
-    pub fn fcvt_r(
-        __value: f64,
-        __ndigit: ::std::os::raw::c_int,
-        __decpt: *mut ::std::os::raw::c_int,
-        __sign: *mut ::std::os::raw::c_int,
-        __buf: *mut ::std::os::raw::c_char,
-        __len: usize,
-    ) -> ::std::os::raw::c_int;
-}
-extern "C" {
-    pub fn qecvt_r(
-        __value: u128,
-        __ndigit: ::std::os::raw::c_int,
-        __decpt: *mut ::std::os::raw::c_int,
-        __sign: *mut ::std::os::raw::c_int,
-        __buf: *mut ::std::os::raw::c_char,
-        __len: usize,
-    ) -> ::std::os::raw::c_int;
-}
-extern "C" {
-    pub fn qfcvt_r(
-        __value: u128,
-        __ndigit: ::std::os::raw::c_int,
-        __decpt: *mut ::std::os::raw::c_int,
-        __sign: *mut ::std::os::raw::c_int,
-        __buf: *mut ::std::os::raw::c_char,
-        __len: usize,
-    ) -> ::std::os::raw::c_int;
-}
-extern "C" {
-    pub fn mblen(__s: *const ::std::os::raw::c_char, __n: usize) -> ::std::os::raw::c_int;
-}
-extern "C" {
-    pub fn mbtowc(
-        __pwc: *mut wchar_t,
-        __s: *const ::std::os::raw::c_char,
-        __n: usize,
-    ) -> ::std::os::raw::c_int;
-}
-extern "C" {
-    pub fn wctomb(__s: *mut ::std::os::raw::c_char, __wchar: wchar_t) -> ::std::os::raw::c_int;
-}
-extern "C" {
-    pub fn mbstowcs(__pwcs: *mut wchar_t, __s: *const ::std::os::raw::c_char, __n: usize) -> usize;
-}
-extern "C" {
-    pub fn wcstombs(__s: *mut ::std::os::raw::c_char, __pwcs: *const wchar_t, __n: usize) -> usize;
-}
-extern "C" {
-    pub fn rpmatch(__response: *const ::std::os::raw::c_char) -> ::std::os::raw::c_int;
-}
-extern "C" {
-    pub fn getsubopt(
-        __optionp: *mut *mut ::std::os::raw::c_char,
-        __tokens: *const *mut ::std::os::raw::c_char,
-        __valuep: *mut *mut ::std::os::raw::c_char,
-    ) -> ::std::os::raw::c_int;
-}
-extern "C" {
-    pub fn getloadavg(__loadavg: *mut f64, __nelem: ::std::os::raw::c_int)
-        -> ::std::os::raw::c_int;
-}
-extern "C" {
-    pub fn memcpy(
-        __dest: *mut ::std::os::raw::c_void,
-        __src: *const ::std::os::raw::c_void,
-        __n: ::std::os::raw::c_ulong,
-    ) -> *mut ::std::os::raw::c_void;
-}
-extern "C" {
-    pub fn memmove(
-        __dest: *mut ::std::os::raw::c_void,
-        __src: *const ::std::os::raw::c_void,
-        __n: ::std::os::raw::c_ulong,
-    ) -> *mut ::std::os::raw::c_void;
-}
-extern "C" {
-    pub fn memccpy(
-        __dest: *mut ::std::os::raw::c_void,
-        __src: *const ::std::os::raw::c_void,
-        __c: ::std::os::raw::c_int,
-        __n: ::std::os::raw::c_ulong,
-    ) -> *mut ::std::os::raw::c_void;
-}
-extern "C" {
-    pub fn memset(
-        __s: *mut ::std::os::raw::c_void,
-        __c: ::std::os::raw::c_int,
-        __n: ::std::os::raw::c_ulong,
-    ) -> *mut ::std::os::raw::c_void;
-}
-extern "C" {
-    pub fn memcmp(
-        __s1: *const ::std::os::raw::c_void,
-        __s2: *const ::std::os::raw::c_void,
-        __n: ::std::os::raw::c_ulong,
-    ) -> ::std::os::raw::c_int;
-}
-extern "C" {
-    pub fn memchr(
-        __s: *const ::std::os::raw::c_void,
-        __c: ::std::os::raw::c_int,
-        __n: ::std::os::raw::c_ulong,
-    ) -> *mut ::std::os::raw::c_void;
-}
-extern "C" {
-    pub fn strcpy(
-        __dest: *mut ::std::os::raw::c_char,
-        __src: *const ::std::os::raw::c_char,
-    ) -> *mut ::std::os::raw::c_char;
-}
-extern "C" {
-    pub fn strncpy(
-        __dest: *mut ::std::os::raw::c_char,
-        __src: *const ::std::os::raw::c_char,
-        __n: ::std::os::raw::c_ulong,
-    ) -> *mut ::std::os::raw::c_char;
-}
-extern "C" {
-    pub fn strcat(
-        __dest: *mut ::std::os::raw::c_char,
-        __src: *const ::std::os::raw::c_char,
-    ) -> *mut ::std::os::raw::c_char;
-}
-extern "C" {
-    pub fn strncat(
-        __dest: *mut ::std::os::raw::c_char,
-        __src: *const ::std::os::raw::c_char,
-        __n: ::std::os::raw::c_ulong,
-    ) -> *mut ::std::os::raw::c_char;
-}
-extern "C" {
-    pub fn strcmp(
-        __s1: *const ::std::os::raw::c_char,
-        __s2: *const ::std::os::raw::c_char,
-    ) -> ::std::os::raw::c_int;
-}
-extern "C" {
-    pub fn strncmp(
-        __s1: *const ::std::os::raw::c_char,
-        __s2: *const ::std::os::raw::c_char,
-        __n: ::std::os::raw::c_ulong,
-    ) -> ::std::os::raw::c_int;
-}
-extern "C" {
-    pub fn strcoll(
-        __s1: *const ::std::os::raw::c_char,
-        __s2: *const ::std::os::raw::c_char,
-    ) -> ::std::os::raw::c_int;
-}
-extern "C" {
-    pub fn strxfrm(
-        __dest: *mut ::std::os::raw::c_char,
-        __src: *const ::std::os::raw::c_char,
-        __n: ::std::os::raw::c_ulong,
-    ) -> ::std::os::raw::c_ulong;
-}
-#[repr(C)]
-#[derive(Debug, Copy, Clone)]
-pub struct __locale_struct {
-    pub __locales: [*mut __locale_data; 13usize],
-    pub __ctype_b: *const ::std::os::raw::c_ushort,
-    pub __ctype_tolower: *const ::std::os::raw::c_int,
-    pub __ctype_toupper: *const ::std::os::raw::c_int,
-    pub __names: [*const ::std::os::raw::c_char; 13usize],
-}
-#[test]
-fn bindgen_test_layout___locale_struct() {
-    assert_eq!(
-        ::std::mem::size_of::<__locale_struct>(),
-        232usize,
-        concat!("Size of: ", stringify!(__locale_struct))
-    );
-    assert_eq!(
-        ::std::mem::align_of::<__locale_struct>(),
-        8usize,
-        concat!("Alignment of ", stringify!(__locale_struct))
-    );
-    assert_eq!(
-        unsafe { &(*(::std::ptr::null::<__locale_struct>())).__locales as *const _ as usize },
-        0usize,
-        concat!(
-            "Offset of field: ",
-            stringify!(__locale_struct),
-            "::",
-            stringify!(__locales)
-        )
-    );
-    assert_eq!(
-        unsafe { &(*(::std::ptr::null::<__locale_struct>())).__ctype_b as *const _ as usize },
-        104usize,
-        concat!(
-            "Offset of field: ",
-            stringify!(__locale_struct),
-            "::",
-            stringify!(__ctype_b)
-        )
-    );
-    assert_eq!(
-        unsafe { &(*(::std::ptr::null::<__locale_struct>())).__ctype_tolower as *const _ as usize },
-        112usize,
-        concat!(
-            "Offset of field: ",
-            stringify!(__locale_struct),
-            "::",
-            stringify!(__ctype_tolower)
-        )
-    );
-    assert_eq!(
-        unsafe { &(*(::std::ptr::null::<__locale_struct>())).__ctype_toupper as *const _ as usize },
-        120usize,
-        concat!(
-            "Offset of field: ",
-            stringify!(__locale_struct),
-            "::",
-            stringify!(__ctype_toupper)
-        )
-    );
-    assert_eq!(
-        unsafe { &(*(::std::ptr::null::<__locale_struct>())).__names as *const _ as usize },
-        128usize,
-        concat!(
-            "Offset of field: ",
-            stringify!(__locale_struct),
-            "::",
-            stringify!(__names)
-        )
-    );
-}
-pub type __locale_t = *mut __locale_struct;
-pub type locale_t = __locale_t;
-extern "C" {
-    pub fn strcoll_l(
-        __s1: *const ::std::os::raw::c_char,
-        __s2: *const ::std::os::raw::c_char,
-        __l: locale_t,
-    ) -> ::std::os::raw::c_int;
-}
-extern "C" {
-    pub fn strxfrm_l(
-        __dest: *mut ::std::os::raw::c_char,
-        __src: *const ::std::os::raw::c_char,
-        __n: usize,
-        __l: locale_t,
-    ) -> usize;
-}
-extern "C" {
-    pub fn strdup(__s: *const ::std::os::raw::c_char) -> *mut ::std::os::raw::c_char;
-}
-extern "C" {
-    pub fn strndup(
-        __string: *const ::std::os::raw::c_char,
-        __n: ::std::os::raw::c_ulong,
-    ) -> *mut ::std::os::raw::c_char;
-}
-extern "C" {
-    pub fn strchr(
-        __s: *const ::std::os::raw::c_char,
-        __c: ::std::os::raw::c_int,
-    ) -> *mut ::std::os::raw::c_char;
-}
-extern "C" {
-    pub fn strrchr(
-        __s: *const ::std::os::raw::c_char,
-        __c: ::std::os::raw::c_int,
-    ) -> *mut ::std::os::raw::c_char;
-}
-extern "C" {
-    pub fn strcspn(
-        __s: *const ::std::os::raw::c_char,
-        __reject: *const ::std::os::raw::c_char,
-    ) -> ::std::os::raw::c_ulong;
-}
-extern "C" {
-    pub fn strspn(
-        __s: *const ::std::os::raw::c_char,
-        __accept: *const ::std::os::raw::c_char,
-    ) -> ::std::os::raw::c_ulong;
-}
-extern "C" {
-    pub fn strpbrk(
-        __s: *const ::std::os::raw::c_char,
-        __accept: *const ::std::os::raw::c_char,
-    ) -> *mut ::std::os::raw::c_char;
-}
-extern "C" {
-    pub fn strstr(
-        __haystack: *const ::std::os::raw::c_char,
-        __needle: *const ::std::os::raw::c_char,
-    ) -> *mut ::std::os::raw::c_char;
-}
-extern "C" {
-    pub fn strtok(
-        __s: *mut ::std::os::raw::c_char,
-        __delim: *const ::std::os::raw::c_char,
-    ) -> *mut ::std::os::raw::c_char;
-}
-extern "C" {
-    pub fn __strtok_r(
-        __s: *mut ::std::os::raw::c_char,
-        __delim: *const ::std::os::raw::c_char,
-        __save_ptr: *mut *mut ::std::os::raw::c_char,
-    ) -> *mut ::std::os::raw::c_char;
-}
-extern "C" {
-    pub fn strtok_r(
-        __s: *mut ::std::os::raw::c_char,
-        __delim: *const ::std::os::raw::c_char,
-        __save_ptr: *mut *mut ::std::os::raw::c_char,
-    ) -> *mut ::std::os::raw::c_char;
-}
-extern "C" {
-    pub fn strlen(__s: *const ::std::os::raw::c_char) -> ::std::os::raw::c_ulong;
-}
-extern "C" {
-    pub fn strnlen(__string: *const ::std::os::raw::c_char, __maxlen: usize) -> usize;
-}
-extern "C" {
-    pub fn strerror(__errnum: ::std::os::raw::c_int) -> *mut ::std::os::raw::c_char;
-}
-extern "C" {
-    #[link_name = "\u{1}__xpg_strerror_r"]
-    pub fn strerror_r(
-        __errnum: ::std::os::raw::c_int,
-        __buf: *mut ::std::os::raw::c_char,
-        __buflen: usize,
-    ) -> ::std::os::raw::c_int;
-}
-extern "C" {
-    pub fn strerror_l(
-        __errnum: ::std::os::raw::c_int,
-        __l: locale_t,
-    ) -> *mut ::std::os::raw::c_char;
-}
-extern "C" {
-    pub fn bcmp(
-        __s1: *const ::std::os::raw::c_void,
-        __s2: *const ::std::os::raw::c_void,
-        __n: ::std::os::raw::c_ulong,
-    ) -> ::std::os::raw::c_int;
-}
-extern "C" {
-    pub fn bcopy(
-        __src: *const ::std::os::raw::c_void,
-        __dest: *mut ::std::os::raw::c_void,
-        __n: usize,
-    );
-}
-extern "C" {
-    pub fn bzero(__s: *mut ::std::os::raw::c_void, __n: ::std::os::raw::c_ulong);
-}
-extern "C" {
-    pub fn index(
-        __s: *const ::std::os::raw::c_char,
-        __c: ::std::os::raw::c_int,
-    ) -> *mut ::std::os::raw::c_char;
-}
-extern "C" {
-    pub fn rindex(
-        __s: *const ::std::os::raw::c_char,
-        __c: ::std::os::raw::c_int,
-    ) -> *mut ::std::os::raw::c_char;
-}
-extern "C" {
-    pub fn ffs(__i: ::std::os::raw::c_int) -> ::std::os::raw::c_int;
-}
-extern "C" {
-    pub fn ffsl(__l: ::std::os::raw::c_long) -> ::std::os::raw::c_int;
-}
-extern "C" {
-    pub fn ffsll(__ll: ::std::os::raw::c_longlong) -> ::std::os::raw::c_int;
-}
-extern "C" {
-    pub fn strcasecmp(
-        __s1: *const ::std::os::raw::c_char,
-        __s2: *const ::std::os::raw::c_char,
-    ) -> ::std::os::raw::c_int;
-}
-extern "C" {
-    pub fn strncasecmp(
-        __s1: *const ::std::os::raw::c_char,
-        __s2: *const ::std::os::raw::c_char,
-        __n: ::std::os::raw::c_ulong,
-    ) -> ::std::os::raw::c_int;
-}
-extern "C" {
-    pub fn strcasecmp_l(
-        __s1: *const ::std::os::raw::c_char,
-        __s2: *const ::std::os::raw::c_char,
-        __loc: locale_t,
-    ) -> ::std::os::raw::c_int;
-}
-extern "C" {
-    pub fn strncasecmp_l(
-        __s1: *const ::std::os::raw::c_char,
-        __s2: *const ::std::os::raw::c_char,
-        __n: usize,
-        __loc: locale_t,
-    ) -> ::std::os::raw::c_int;
-}
-extern "C" {
-    pub fn explicit_bzero(__s: *mut ::std::os::raw::c_void, __n: usize);
-}
-extern "C" {
-    pub fn strsep(
-        __stringp: *mut *mut ::std::os::raw::c_char,
-        __delim: *const ::std::os::raw::c_char,
-    ) -> *mut ::std::os::raw::c_char;
-}
-extern "C" {
-    pub fn strsignal(__sig: ::std::os::raw::c_int) -> *mut ::std::os::raw::c_char;
-}
-extern "C" {
-    pub fn __stpcpy(
-        __dest: *mut ::std::os::raw::c_char,
-        __src: *const ::std::os::raw::c_char,
-    ) -> *mut ::std::os::raw::c_char;
-}
-extern "C" {
-    pub fn stpcpy(
-        __dest: *mut ::std::os::raw::c_char,
-        __src: *const ::std::os::raw::c_char,
-    ) -> *mut ::std::os::raw::c_char;
-}
-extern "C" {
-    pub fn __stpncpy(
-        __dest: *mut ::std::os::raw::c_char,
-        __src: *const ::std::os::raw::c_char,
-        __n: usize,
-    ) -> *mut ::std::os::raw::c_char;
-}
-extern "C" {
-    pub fn stpncpy(
-        __dest: *mut ::std::os::raw::c_char,
-        __src: *const ::std::os::raw::c_char,
-        __n: ::std::os::raw::c_ulong,
-    ) -> *mut ::std::os::raw::c_char;
-}
-impl TF_TString_Type {
-    pub const TF_TSTR_TYPE_MASK: TF_TString_Type = TF_TString_Type::TF_TSTR_VIEW;
-}
-#[repr(u32)]
-#[derive(Debug, Copy, Clone, PartialEq, Eq, Hash)]
-pub enum TF_TString_Type {
-    TF_TSTR_SMALL = 0,
-    TF_TSTR_LARGE = 1,
-    TF_TSTR_OFFSET = 2,
-    TF_TSTR_VIEW = 3,
-}
-#[repr(C)]
-#[derive(Debug, Copy, Clone)]
-pub struct TF_TString_Large {
-    pub size: usize,
-    pub cap: usize,
-    pub ptr: *mut ::std::os::raw::c_char,
-}
-#[test]
-fn bindgen_test_layout_TF_TString_Large() {
-    assert_eq!(
-        ::std::mem::size_of::<TF_TString_Large>(),
-        24usize,
-        concat!("Size of: ", stringify!(TF_TString_Large))
-    );
-    assert_eq!(
-        ::std::mem::align_of::<TF_TString_Large>(),
-        8usize,
-        concat!("Alignment of ", stringify!(TF_TString_Large))
-    );
-    assert_eq!(
-        unsafe { &(*(::std::ptr::null::<TF_TString_Large>())).size as *const _ as usize },
-        0usize,
-        concat!(
-            "Offset of field: ",
-            stringify!(TF_TString_Large),
-            "::",
-            stringify!(size)
-        )
-    );
-    assert_eq!(
-        unsafe { &(*(::std::ptr::null::<TF_TString_Large>())).cap as *const _ as usize },
-        8usize,
-        concat!(
-            "Offset of field: ",
-            stringify!(TF_TString_Large),
-            "::",
-            stringify!(cap)
-        )
-    );
-    assert_eq!(
-        unsafe { &(*(::std::ptr::null::<TF_TString_Large>())).ptr as *const _ as usize },
-        16usize,
-        concat!(
-            "Offset of field: ",
-            stringify!(TF_TString_Large),
-            "::",
-            stringify!(ptr)
-        )
-    );
-}
-#[repr(C)]
-#[derive(Debug, Copy, Clone)]
-pub struct TF_TString_Offset {
-    pub size: u32,
-    pub offset: u32,
-    pub count: u32,
-}
-#[test]
-fn bindgen_test_layout_TF_TString_Offset() {
-    assert_eq!(
-        ::std::mem::size_of::<TF_TString_Offset>(),
-        12usize,
-        concat!("Size of: ", stringify!(TF_TString_Offset))
-    );
-    assert_eq!(
-        ::std::mem::align_of::<TF_TString_Offset>(),
-        4usize,
-        concat!("Alignment of ", stringify!(TF_TString_Offset))
-    );
-    assert_eq!(
-        unsafe { &(*(::std::ptr::null::<TF_TString_Offset>())).size as *const _ as usize },
-        0usize,
-        concat!(
-            "Offset of field: ",
-            stringify!(TF_TString_Offset),
-            "::",
-            stringify!(size)
-        )
-    );
-    assert_eq!(
-        unsafe { &(*(::std::ptr::null::<TF_TString_Offset>())).offset as *const _ as usize },
-        4usize,
-        concat!(
-            "Offset of field: ",
-            stringify!(TF_TString_Offset),
-            "::",
-            stringify!(offset)
-        )
-    );
-    assert_eq!(
-        unsafe { &(*(::std::ptr::null::<TF_TString_Offset>())).count as *const _ as usize },
-        8usize,
-        concat!(
-            "Offset of field: ",
-            stringify!(TF_TString_Offset),
-            "::",
-            stringify!(count)
-        )
-    );
-}
-#[repr(C)]
-#[derive(Debug, Copy, Clone)]
-pub struct TF_TString_View {
-    pub size: usize,
-    pub ptr: *const ::std::os::raw::c_char,
-}
-#[test]
-fn bindgen_test_layout_TF_TString_View() {
-    assert_eq!(
-        ::std::mem::size_of::<TF_TString_View>(),
-        16usize,
-        concat!("Size of: ", stringify!(TF_TString_View))
-    );
-    assert_eq!(
-        ::std::mem::align_of::<TF_TString_View>(),
-        8usize,
-        concat!("Alignment of ", stringify!(TF_TString_View))
-    );
-    assert_eq!(
-        unsafe { &(*(::std::ptr::null::<TF_TString_View>())).size as *const _ as usize },
-        0usize,
-        concat!(
-            "Offset of field: ",
-            stringify!(TF_TString_View),
-            "::",
-            stringify!(size)
-        )
-    );
-    assert_eq!(
-        unsafe { &(*(::std::ptr::null::<TF_TString_View>())).ptr as *const _ as usize },
-        8usize,
-        concat!(
-            "Offset of field: ",
-            stringify!(TF_TString_View),
-            "::",
-            stringify!(ptr)
-        )
-    );
-}
-#[repr(C)]
-#[derive(Debug, Copy, Clone)]
-pub struct TF_TString_Raw {
-    pub raw: [u8; 24usize],
-}
-#[test]
-fn bindgen_test_layout_TF_TString_Raw() {
-    assert_eq!(
-        ::std::mem::size_of::<TF_TString_Raw>(),
-        24usize,
-        concat!("Size of: ", stringify!(TF_TString_Raw))
-    );
-    assert_eq!(
-        ::std::mem::align_of::<TF_TString_Raw>(),
-        1usize,
-        concat!("Alignment of ", stringify!(TF_TString_Raw))
-    );
-    assert_eq!(
-        unsafe { &(*(::std::ptr::null::<TF_TString_Raw>())).raw as *const _ as usize },
-        0usize,
-        concat!(
-            "Offset of field: ",
-            stringify!(TF_TString_Raw),
-            "::",
-            stringify!(raw)
-        )
-    );
-}
-#[repr(C)]
-#[derive(Copy, Clone)]
-pub union TF_TString_Union {
-    pub large: TF_TString_Large,
-    pub offset: TF_TString_Offset,
-    pub view: TF_TString_View,
-    pub raw: TF_TString_Raw,
-    _bindgen_union_align: [u64; 3usize],
-}
-#[test]
-fn bindgen_test_layout_TF_TString_Union() {
-    assert_eq!(
-        ::std::mem::size_of::<TF_TString_Union>(),
-        24usize,
-        concat!("Size of: ", stringify!(TF_TString_Union))
-    );
-    assert_eq!(
-        ::std::mem::align_of::<TF_TString_Union>(),
-        8usize,
-        concat!("Alignment of ", stringify!(TF_TString_Union))
-    );
-    assert_eq!(
-        unsafe { &(*(::std::ptr::null::<TF_TString_Union>())).large as *const _ as usize },
-        0usize,
-        concat!(
-            "Offset of field: ",
-            stringify!(TF_TString_Union),
-            "::",
-            stringify!(large)
-        )
-    );
-    assert_eq!(
-        unsafe { &(*(::std::ptr::null::<TF_TString_Union>())).offset as *const _ as usize },
-        0usize,
-        concat!(
-            "Offset of field: ",
-            stringify!(TF_TString_Union),
-            "::",
-            stringify!(offset)
-        )
-    );
-    assert_eq!(
-        unsafe { &(*(::std::ptr::null::<TF_TString_Union>())).view as *const _ as usize },
-        0usize,
-        concat!(
-            "Offset of field: ",
-            stringify!(TF_TString_Union),
-            "::",
-            stringify!(view)
-        )
-    );
-    assert_eq!(
-        unsafe { &(*(::std::ptr::null::<TF_TString_Union>())).raw as *const _ as usize },
-        0usize,
-        concat!(
-            "Offset of field: ",
-            stringify!(TF_TString_Union),
-            "::",
-            stringify!(raw)
-        )
-    );
-}
-pub const TF_TString_SmallCapacity: _bindgen_ty_1 = _bindgen_ty_1::TF_TString_SmallCapacity;
-#[repr(u32)]
-#[derive(Debug, Copy, Clone, PartialEq, Eq, Hash)]
-pub enum _bindgen_ty_1 {
-    TF_TString_SmallCapacity = 22,
-}
-#[repr(C)]
-#[derive(Debug, Copy, Clone)]
-pub struct TF_TString_Small {
-    pub size: u8,
-    pub str_: [::std::os::raw::c_char; 23usize],
-}
-#[test]
-fn bindgen_test_layout_TF_TString_Small() {
-    assert_eq!(
-        ::std::mem::size_of::<TF_TString_Small>(),
-        24usize,
-        concat!("Size of: ", stringify!(TF_TString_Small))
-    );
-    assert_eq!(
-        ::std::mem::align_of::<TF_TString_Small>(),
-        1usize,
-        concat!("Alignment of ", stringify!(TF_TString_Small))
-    );
-    assert_eq!(
-        unsafe { &(*(::std::ptr::null::<TF_TString_Small>())).size as *const _ as usize },
-        0usize,
-        concat!(
-            "Offset of field: ",
-            stringify!(TF_TString_Small),
-            "::",
-            stringify!(size)
-        )
-    );
-    assert_eq!(
-        unsafe { &(*(::std::ptr::null::<TF_TString_Small>())).str_ as *const _ as usize },
-        1usize,
-        concat!(
-            "Offset of field: ",
-            stringify!(TF_TString_Small),
-            "::",
-            stringify!(str_)
-        )
-    );
-}
-#[repr(C)]
-#[derive(Copy, Clone)]
-pub struct TF_TString {
-    pub u: TF_TString__bindgen_ty_1,
-}
-#[repr(C)]
-#[derive(Copy, Clone)]
-pub union TF_TString__bindgen_ty_1 {
-    pub smll: TF_TString_Small,
-    pub large: TF_TString_Large,
-    pub offset: TF_TString_Offset,
-    pub view: TF_TString_View,
-    pub raw: TF_TString_Raw,
-    _bindgen_union_align: [u64; 3usize],
-}
-#[test]
-fn bindgen_test_layout_TF_TString__bindgen_ty_1() {
-    assert_eq!(
-        ::std::mem::size_of::<TF_TString__bindgen_ty_1>(),
-        24usize,
-        concat!("Size of: ", stringify!(TF_TString__bindgen_ty_1))
-    );
-    assert_eq!(
-        ::std::mem::align_of::<TF_TString__bindgen_ty_1>(),
-        8usize,
-        concat!("Alignment of ", stringify!(TF_TString__bindgen_ty_1))
-    );
-    assert_eq!(
-        unsafe { &(*(::std::ptr::null::<TF_TString__bindgen_ty_1>())).smll as *const _ as usize },
-        0usize,
-        concat!(
-            "Offset of field: ",
-            stringify!(TF_TString__bindgen_ty_1),
-            "::",
-            stringify!(smll)
-        )
-    );
-    assert_eq!(
-        unsafe { &(*(::std::ptr::null::<TF_TString__bindgen_ty_1>())).large as *const _ as usize },
-        0usize,
-        concat!(
-            "Offset of field: ",
-            stringify!(TF_TString__bindgen_ty_1),
-            "::",
-            stringify!(large)
-        )
-    );
-    assert_eq!(
-        unsafe { &(*(::std::ptr::null::<TF_TString__bindgen_ty_1>())).offset as *const _ as usize },
-        0usize,
-        concat!(
-            "Offset of field: ",
-            stringify!(TF_TString__bindgen_ty_1),
-            "::",
-            stringify!(offset)
-        )
-    );
-    assert_eq!(
-        unsafe { &(*(::std::ptr::null::<TF_TString__bindgen_ty_1>())).view as *const _ as usize },
-        0usize,
-        concat!(
-            "Offset of field: ",
-            stringify!(TF_TString__bindgen_ty_1),
-            "::",
-            stringify!(view)
-        )
-    );
-    assert_eq!(
-        unsafe { &(*(::std::ptr::null::<TF_TString__bindgen_ty_1>())).raw as *const _ as usize },
-        0usize,
-        concat!(
-            "Offset of field: ",
-            stringify!(TF_TString__bindgen_ty_1),
-            "::",
-            stringify!(raw)
-        )
-    );
-}
-#[test]
-fn bindgen_test_layout_TF_TString() {
-    assert_eq!(
-        ::std::mem::size_of::<TF_TString>(),
-        24usize,
-        concat!("Size of: ", stringify!(TF_TString))
-    );
-    assert_eq!(
-        ::std::mem::align_of::<TF_TString>(),
-        8usize,
-        concat!("Alignment of ", stringify!(TF_TString))
-    );
-    assert_eq!(
-        unsafe { &(*(::std::ptr::null::<TF_TString>())).u as *const _ as usize },
-        0usize,
-        concat!(
-            "Offset of field: ",
-            stringify!(TF_TString),
-            "::",
-            stringify!(u)
-        )
-    );
-}
-extern "C" {
-    pub fn TF_StringInit(t: *mut TF_TString);
-}
-extern "C" {
-    pub fn TF_StringCopy(dst: *mut TF_TString, src: *const ::std::os::raw::c_char, size: usize);
-}
-extern "C" {
-    pub fn TF_StringAssignView(
-        dst: *mut TF_TString,
-        src: *const ::std::os::raw::c_char,
-        size: usize,
-    );
-}
-extern "C" {
-    pub fn TF_StringGetDataPointer(tstr: *const TF_TString) -> *const ::std::os::raw::c_char;
-}
-extern "C" {
-    pub fn TF_StringGetType(str_: *const TF_TString) -> TF_TString_Type;
-}
-extern "C" {
-    pub fn TF_StringGetSize(tstr: *const TF_TString) -> usize;
-}
-extern "C" {
-    pub fn TF_StringGetCapacity(str_: *const TF_TString) -> usize;
-}
-extern "C" {
-    pub fn TF_StringDealloc(tstr: *mut TF_TString);
-}
-extern "C" {
-    pub fn TF_Version() -> *const ::std::os::raw::c_char;
-}
-#[repr(C)]
-#[derive(Debug, Copy, Clone)]
-pub struct TF_Buffer {
-    pub data: *const ::std::os::raw::c_void,
-    pub length: usize,
-    pub data_deallocator: ::std::option::Option<
-        unsafe extern "C" fn(data: *mut ::std::os::raw::c_void, length: usize),
-    >,
-}
-#[test]
-fn bindgen_test_layout_TF_Buffer() {
-    assert_eq!(
-        ::std::mem::size_of::<TF_Buffer>(),
-        24usize,
-        concat!("Size of: ", stringify!(TF_Buffer))
-    );
-    assert_eq!(
-        ::std::mem::align_of::<TF_Buffer>(),
-        8usize,
-        concat!("Alignment of ", stringify!(TF_Buffer))
-    );
-    assert_eq!(
-        unsafe { &(*(::std::ptr::null::<TF_Buffer>())).data as *const _ as usize },
-        0usize,
-        concat!(
-            "Offset of field: ",
-            stringify!(TF_Buffer),
-            "::",
-            stringify!(data)
-        )
-    );
-    assert_eq!(
-        unsafe { &(*(::std::ptr::null::<TF_Buffer>())).length as *const _ as usize },
-        8usize,
-        concat!(
-            "Offset of field: ",
-            stringify!(TF_Buffer),
-            "::",
-            stringify!(length)
-        )
-    );
-    assert_eq!(
-        unsafe { &(*(::std::ptr::null::<TF_Buffer>())).data_deallocator as *const _ as usize },
-        16usize,
-        concat!(
-            "Offset of field: ",
-            stringify!(TF_Buffer),
-            "::",
-            stringify!(data_deallocator)
-        )
-    );
-}
-extern "C" {
-    pub fn TF_NewBufferFromString(
-        proto: *const ::std::os::raw::c_void,
-        proto_len: usize,
-    ) -> *mut TF_Buffer;
-}
-extern "C" {
-    pub fn TF_NewBuffer() -> *mut TF_Buffer;
-}
-extern "C" {
-    pub fn TF_DeleteBuffer(arg1: *mut TF_Buffer);
-}
-extern "C" {
-    pub fn TF_GetBuffer(buffer: *mut TF_Buffer) -> TF_Buffer;
-}
-#[repr(C)]
-#[derive(Debug, Copy, Clone)]
-pub struct TF_StringView {
-    pub data: *const ::std::os::raw::c_char,
-    pub len: usize,
-}
-#[test]
-fn bindgen_test_layout_TF_StringView() {
-    assert_eq!(
-        ::std::mem::size_of::<TF_StringView>(),
-        16usize,
-        concat!("Size of: ", stringify!(TF_StringView))
-    );
-    assert_eq!(
-        ::std::mem::align_of::<TF_StringView>(),
-        8usize,
-        concat!("Alignment of ", stringify!(TF_StringView))
-    );
-    assert_eq!(
-        unsafe { &(*(::std::ptr::null::<TF_StringView>())).data as *const _ as usize },
-        0usize,
-        concat!(
-            "Offset of field: ",
-            stringify!(TF_StringView),
-            "::",
-            stringify!(data)
-        )
-    );
-    assert_eq!(
-        unsafe { &(*(::std::ptr::null::<TF_StringView>())).len as *const _ as usize },
-        8usize,
-        concat!(
-            "Offset of field: ",
-            stringify!(TF_StringView),
-            "::",
-            stringify!(len)
-        )
-    );
-}
-#[repr(C)]
-#[derive(Debug, Copy, Clone)]
-pub struct TF_SessionOptions {
-    _unused: [u8; 0],
-}
-extern "C" {
-    pub fn TF_NewSessionOptions() -> *mut TF_SessionOptions;
-}
-extern "C" {
-    pub fn TF_SetTarget(options: *mut TF_SessionOptions, target: *const ::std::os::raw::c_char);
-}
-extern "C" {
-    pub fn TF_SetConfig(
-        options: *mut TF_SessionOptions,
-        proto: *const ::std::os::raw::c_void,
-        proto_len: usize,
-        status: *mut TF_Status,
-    );
-}
-extern "C" {
-    pub fn TF_DeleteSessionOptions(arg1: *mut TF_SessionOptions);
-}
-#[repr(C)]
-#[derive(Debug, Copy, Clone)]
-pub struct TF_Graph {
-    _unused: [u8; 0],
-}
-extern "C" {
-    pub fn TF_NewGraph() -> *mut TF_Graph;
-}
-extern "C" {
-    pub fn TF_DeleteGraph(arg1: *mut TF_Graph);
-}
-#[repr(C)]
-#[derive(Debug, Copy, Clone)]
-pub struct TF_OperationDescription {
-    _unused: [u8; 0],
-}
-#[repr(C)]
-#[derive(Debug, Copy, Clone)]
-pub struct TF_Operation {
-    _unused: [u8; 0],
-}
-#[repr(C)]
-#[derive(Debug, Copy, Clone)]
-pub struct TF_Input {
-    pub oper: *mut TF_Operation,
-    pub index: ::std::os::raw::c_int,
-}
-#[test]
-fn bindgen_test_layout_TF_Input() {
-    assert_eq!(
-        ::std::mem::size_of::<TF_Input>(),
-        16usize,
-        concat!("Size of: ", stringify!(TF_Input))
-    );
-    assert_eq!(
-        ::std::mem::align_of::<TF_Input>(),
-        8usize,
-        concat!("Alignment of ", stringify!(TF_Input))
-    );
-    assert_eq!(
-        unsafe { &(*(::std::ptr::null::<TF_Input>())).oper as *const _ as usize },
-        0usize,
-        concat!(
-            "Offset of field: ",
-            stringify!(TF_Input),
-            "::",
-            stringify!(oper)
-        )
-    );
-    assert_eq!(
-        unsafe { &(*(::std::ptr::null::<TF_Input>())).index as *const _ as usize },
-        8usize,
-        concat!(
-            "Offset of field: ",
-            stringify!(TF_Input),
-            "::",
-            stringify!(index)
-        )
-    );
-}
-#[repr(C)]
-#[derive(Debug, Copy, Clone)]
-pub struct TF_Output {
-    pub oper: *mut TF_Operation,
-    pub index: ::std::os::raw::c_int,
-}
-#[test]
-fn bindgen_test_layout_TF_Output() {
-    assert_eq!(
-        ::std::mem::size_of::<TF_Output>(),
-        16usize,
-        concat!("Size of: ", stringify!(TF_Output))
-    );
-    assert_eq!(
-        ::std::mem::align_of::<TF_Output>(),
-        8usize,
-        concat!("Alignment of ", stringify!(TF_Output))
-    );
-    assert_eq!(
-        unsafe { &(*(::std::ptr::null::<TF_Output>())).oper as *const _ as usize },
-        0usize,
-        concat!(
-            "Offset of field: ",
-            stringify!(TF_Output),
-            "::",
-            stringify!(oper)
-        )
-    );
-    assert_eq!(
-        unsafe { &(*(::std::ptr::null::<TF_Output>())).index as *const _ as usize },
-        8usize,
-        concat!(
-            "Offset of field: ",
-            stringify!(TF_Output),
-            "::",
-            stringify!(index)
-        )
-    );
-}
-#[repr(C)]
-#[derive(Debug, Copy, Clone)]
-pub struct TF_Function {
-    _unused: [u8; 0],
-}
-#[repr(C)]
-#[derive(Debug, Copy, Clone)]
-pub struct TF_FunctionOptions {
-    _unused: [u8; 0],
-}
-extern "C" {
-    pub fn TF_GraphSetTensorShape(
-        graph: *mut TF_Graph,
-        output: TF_Output,
-        dims: *const i64,
-        num_dims: ::std::os::raw::c_int,
-        status: *mut TF_Status,
-    );
-}
-extern "C" {
-    pub fn TF_GraphGetTensorNumDims(
-        graph: *mut TF_Graph,
-        output: TF_Output,
-        status: *mut TF_Status,
-    ) -> ::std::os::raw::c_int;
-}
-extern "C" {
-    pub fn TF_GraphGetTensorShape(
-        graph: *mut TF_Graph,
-        output: TF_Output,
-        dims: *mut i64,
-        num_dims: ::std::os::raw::c_int,
-        status: *mut TF_Status,
-    );
-}
-extern "C" {
-    pub fn TF_NewOperation(
-        graph: *mut TF_Graph,
-        op_type: *const ::std::os::raw::c_char,
-        oper_name: *const ::std::os::raw::c_char,
-    ) -> *mut TF_OperationDescription;
-}
-extern "C" {
-    pub fn TF_SetDevice(desc: *mut TF_OperationDescription, device: *const ::std::os::raw::c_char);
-}
-extern "C" {
-    pub fn TF_AddInput(desc: *mut TF_OperationDescription, input: TF_Output);
-}
-extern "C" {
-    pub fn TF_AddInputList(
-        desc: *mut TF_OperationDescription,
-        inputs: *const TF_Output,
-        num_inputs: ::std::os::raw::c_int,
-    );
-}
-extern "C" {
-    pub fn TF_AddControlInput(desc: *mut TF_OperationDescription, input: *mut TF_Operation);
-}
-extern "C" {
-    pub fn TF_ColocateWith(desc: *mut TF_OperationDescription, op: *mut TF_Operation);
-}
-extern "C" {
-    pub fn TF_SetAttrString(
-        desc: *mut TF_OperationDescription,
-        attr_name: *const ::std::os::raw::c_char,
-        value: *const ::std::os::raw::c_void,
-        length: usize,
-    );
-}
-extern "C" {
-    pub fn TF_SetAttrStringList(
-        desc: *mut TF_OperationDescription,
-        attr_name: *const ::std::os::raw::c_char,
-        values: *const *const ::std::os::raw::c_void,
-        lengths: *const usize,
-        num_values: ::std::os::raw::c_int,
-    );
-}
-extern "C" {
-    pub fn TF_SetAttrInt(
-        desc: *mut TF_OperationDescription,
-        attr_name: *const ::std::os::raw::c_char,
-        value: i64,
-    );
-}
-extern "C" {
-    pub fn TF_SetAttrIntList(
-        desc: *mut TF_OperationDescription,
-        attr_name: *const ::std::os::raw::c_char,
-        values: *const i64,
-        num_values: ::std::os::raw::c_int,
-    );
-}
-extern "C" {
-    pub fn TF_SetAttrFloat(
-        desc: *mut TF_OperationDescription,
-        attr_name: *const ::std::os::raw::c_char,
-        value: f32,
-    );
-}
-extern "C" {
-    pub fn TF_SetAttrFloatList(
-        desc: *mut TF_OperationDescription,
-        attr_name: *const ::std::os::raw::c_char,
-        values: *const f32,
-        num_values: ::std::os::raw::c_int,
-    );
-}
-extern "C" {
-    pub fn TF_SetAttrBool(
-        desc: *mut TF_OperationDescription,
-        attr_name: *const ::std::os::raw::c_char,
-        value: ::std::os::raw::c_uchar,
-    );
-}
-extern "C" {
-    pub fn TF_SetAttrBoolList(
-        desc: *mut TF_OperationDescription,
-        attr_name: *const ::std::os::raw::c_char,
-        values: *const ::std::os::raw::c_uchar,
-        num_values: ::std::os::raw::c_int,
-    );
-}
-extern "C" {
-    pub fn TF_SetAttrType(
-        desc: *mut TF_OperationDescription,
-        attr_name: *const ::std::os::raw::c_char,
-        value: TF_DataType,
-    );
-}
-extern "C" {
-    pub fn TF_SetAttrTypeList(
-        desc: *mut TF_OperationDescription,
-        attr_name: *const ::std::os::raw::c_char,
-        values: *const TF_DataType,
-        num_values: ::std::os::raw::c_int,
-    );
-}
-extern "C" {
-    pub fn TF_SetAttrPlaceholder(
-        desc: *mut TF_OperationDescription,
-        attr_name: *const ::std::os::raw::c_char,
-        placeholder: *const ::std::os::raw::c_char,
-    );
-}
-extern "C" {
-    pub fn TF_SetAttrFuncName(
-        desc: *mut TF_OperationDescription,
-        attr_name: *const ::std::os::raw::c_char,
-        value: *const ::std::os::raw::c_char,
-        length: usize,
-    );
-}
-extern "C" {
-    pub fn TF_SetAttrShape(
-        desc: *mut TF_OperationDescription,
-        attr_name: *const ::std::os::raw::c_char,
-        dims: *const i64,
-        num_dims: ::std::os::raw::c_int,
-    );
-}
-extern "C" {
-    pub fn TF_SetAttrShapeList(
-        desc: *mut TF_OperationDescription,
-        attr_name: *const ::std::os::raw::c_char,
-        dims: *const *const i64,
-        num_dims: *const ::std::os::raw::c_int,
-        num_shapes: ::std::os::raw::c_int,
-    );
-}
-extern "C" {
-    pub fn TF_SetAttrTensorShapeProto(
-        desc: *mut TF_OperationDescription,
-        attr_name: *const ::std::os::raw::c_char,
-        proto: *const ::std::os::raw::c_void,
-        proto_len: usize,
-        status: *mut TF_Status,
-    );
-}
-extern "C" {
-    pub fn TF_SetAttrTensorShapeProtoList(
-        desc: *mut TF_OperationDescription,
-        attr_name: *const ::std::os::raw::c_char,
-        protos: *const *const ::std::os::raw::c_void,
-        proto_lens: *const usize,
-        num_shapes: ::std::os::raw::c_int,
-        status: *mut TF_Status,
-    );
-}
-extern "C" {
-    pub fn TF_SetAttrTensor(
-        desc: *mut TF_OperationDescription,
-        attr_name: *const ::std::os::raw::c_char,
-        value: *mut TF_Tensor,
-        status: *mut TF_Status,
-    );
-}
-extern "C" {
-    pub fn TF_SetAttrTensorList(
-        desc: *mut TF_OperationDescription,
-        attr_name: *const ::std::os::raw::c_char,
-        values: *const *mut TF_Tensor,
-        num_values: ::std::os::raw::c_int,
-        status: *mut TF_Status,
-    );
-}
-extern "C" {
-    pub fn TF_SetAttrValueProto(
-        desc: *mut TF_OperationDescription,
-        attr_name: *const ::std::os::raw::c_char,
-        proto: *const ::std::os::raw::c_void,
-        proto_len: usize,
-        status: *mut TF_Status,
-    );
-}
-extern "C" {
-    pub fn TF_FinishOperation(
-        desc: *mut TF_OperationDescription,
-        status: *mut TF_Status,
-    ) -> *mut TF_Operation;
-}
-extern "C" {
-    pub fn TF_OperationName(oper: *mut TF_Operation) -> *const ::std::os::raw::c_char;
-}
-extern "C" {
-    pub fn TF_OperationOpType(oper: *mut TF_Operation) -> *const ::std::os::raw::c_char;
-}
-extern "C" {
-    pub fn TF_OperationDevice(oper: *mut TF_Operation) -> *const ::std::os::raw::c_char;
-}
-extern "C" {
-    pub fn TF_OperationNumOutputs(oper: *mut TF_Operation) -> ::std::os::raw::c_int;
-}
-extern "C" {
-    pub fn TF_OperationOutputType(oper_out: TF_Output) -> TF_DataType;
-}
-extern "C" {
-    pub fn TF_OperationOutputListLength(
-        oper: *mut TF_Operation,
-        arg_name: *const ::std::os::raw::c_char,
-        status: *mut TF_Status,
-    ) -> ::std::os::raw::c_int;
-}
-extern "C" {
-    pub fn TF_OperationNumInputs(oper: *mut TF_Operation) -> ::std::os::raw::c_int;
-}
-extern "C" {
-    pub fn TF_OperationInputType(oper_in: TF_Input) -> TF_DataType;
-}
-extern "C" {
-    pub fn TF_OperationInputListLength(
-        oper: *mut TF_Operation,
-        arg_name: *const ::std::os::raw::c_char,
-        status: *mut TF_Status,
-    ) -> ::std::os::raw::c_int;
-}
-extern "C" {
-    pub fn TF_OperationInput(oper_in: TF_Input) -> TF_Output;
-}
-extern "C" {
-    pub fn TF_OperationAllInputs(
-        oper: *mut TF_Operation,
-        inputs: *mut TF_Output,
-        max_inputs: ::std::os::raw::c_int,
-    );
-}
-extern "C" {
-    pub fn TF_OperationOutputNumConsumers(oper_out: TF_Output) -> ::std::os::raw::c_int;
-}
-extern "C" {
-    pub fn TF_OperationOutputConsumers(
-        oper_out: TF_Output,
-        consumers: *mut TF_Input,
-        max_consumers: ::std::os::raw::c_int,
-    ) -> ::std::os::raw::c_int;
-}
-extern "C" {
-    pub fn TF_OperationNumControlInputs(oper: *mut TF_Operation) -> ::std::os::raw::c_int;
-}
-extern "C" {
-    pub fn TF_OperationGetControlInputs(
-        oper: *mut TF_Operation,
-        control_inputs: *mut *mut TF_Operation,
-        max_control_inputs: ::std::os::raw::c_int,
-    ) -> ::std::os::raw::c_int;
-}
-extern "C" {
-    pub fn TF_OperationNumControlOutputs(oper: *mut TF_Operation) -> ::std::os::raw::c_int;
-}
-extern "C" {
-    pub fn TF_OperationGetControlOutputs(
-        oper: *mut TF_Operation,
-        control_outputs: *mut *mut TF_Operation,
-        max_control_outputs: ::std::os::raw::c_int,
-    ) -> ::std::os::raw::c_int;
-}
-#[repr(C)]
-#[derive(Debug, Copy, Clone)]
-pub struct TF_AttrMetadata {
-    pub is_list: ::std::os::raw::c_uchar,
-    pub list_size: i64,
-    pub type_: TF_AttrType,
-    pub total_size: i64,
-}
-#[test]
-fn bindgen_test_layout_TF_AttrMetadata() {
-    assert_eq!(
-        ::std::mem::size_of::<TF_AttrMetadata>(),
-        32usize,
-        concat!("Size of: ", stringify!(TF_AttrMetadata))
-    );
-    assert_eq!(
-        ::std::mem::align_of::<TF_AttrMetadata>(),
-        8usize,
-        concat!("Alignment of ", stringify!(TF_AttrMetadata))
-    );
-    assert_eq!(
-        unsafe { &(*(::std::ptr::null::<TF_AttrMetadata>())).is_list as *const _ as usize },
-        0usize,
-        concat!(
-            "Offset of field: ",
-            stringify!(TF_AttrMetadata),
-            "::",
-            stringify!(is_list)
-        )
-    );
-    assert_eq!(
-        unsafe { &(*(::std::ptr::null::<TF_AttrMetadata>())).list_size as *const _ as usize },
-        8usize,
-        concat!(
-            "Offset of field: ",
-            stringify!(TF_AttrMetadata),
-            "::",
-            stringify!(list_size)
-        )
-    );
-    assert_eq!(
-        unsafe { &(*(::std::ptr::null::<TF_AttrMetadata>())).type_ as *const _ as usize },
-        16usize,
-        concat!(
-            "Offset of field: ",
-            stringify!(TF_AttrMetadata),
-            "::",
-            stringify!(type_)
-        )
-    );
-    assert_eq!(
-        unsafe { &(*(::std::ptr::null::<TF_AttrMetadata>())).total_size as *const _ as usize },
-        24usize,
-        concat!(
-            "Offset of field: ",
-            stringify!(TF_AttrMetadata),
-            "::",
-            stringify!(total_size)
-        )
-    );
-}
-extern "C" {
-    pub fn TF_OperationGetAttrMetadata(
-        oper: *mut TF_Operation,
-        attr_name: *const ::std::os::raw::c_char,
-        status: *mut TF_Status,
-    ) -> TF_AttrMetadata;
-}
-extern "C" {
-    pub fn TF_OperationGetAttrString(
-        oper: *mut TF_Operation,
-        attr_name: *const ::std::os::raw::c_char,
-        value: *mut ::std::os::raw::c_void,
-        max_length: usize,
-        status: *mut TF_Status,
-    );
-}
-extern "C" {
-    pub fn TF_OperationGetAttrStringList(
-        oper: *mut TF_Operation,
-        attr_name: *const ::std::os::raw::c_char,
-        values: *mut *mut ::std::os::raw::c_void,
-        lengths: *mut usize,
-        max_values: ::std::os::raw::c_int,
-        storage: *mut ::std::os::raw::c_void,
-        storage_size: usize,
-        status: *mut TF_Status,
-    );
-}
-extern "C" {
-    pub fn TF_OperationGetAttrInt(
-        oper: *mut TF_Operation,
-        attr_name: *const ::std::os::raw::c_char,
-        value: *mut i64,
-        status: *mut TF_Status,
-    );
-}
-extern "C" {
-    pub fn TF_OperationGetAttrIntList(
-        oper: *mut TF_Operation,
-        attr_name: *const ::std::os::raw::c_char,
-        values: *mut i64,
-        max_values: ::std::os::raw::c_int,
-        status: *mut TF_Status,
-    );
-}
-extern "C" {
-    pub fn TF_OperationGetAttrFloat(
-        oper: *mut TF_Operation,
-        attr_name: *const ::std::os::raw::c_char,
-        value: *mut f32,
-        status: *mut TF_Status,
-    );
-}
-extern "C" {
-    pub fn TF_OperationGetAttrFloatList(
-        oper: *mut TF_Operation,
-        attr_name: *const ::std::os::raw::c_char,
-        values: *mut f32,
-        max_values: ::std::os::raw::c_int,
-        status: *mut TF_Status,
-    );
-}
-extern "C" {
-    pub fn TF_OperationGetAttrBool(
-        oper: *mut TF_Operation,
-        attr_name: *const ::std::os::raw::c_char,
-        value: *mut ::std::os::raw::c_uchar,
-        status: *mut TF_Status,
-    );
-}
-extern "C" {
-    pub fn TF_OperationGetAttrBoolList(
-        oper: *mut TF_Operation,
-        attr_name: *const ::std::os::raw::c_char,
-        values: *mut ::std::os::raw::c_uchar,
-        max_values: ::std::os::raw::c_int,
-        status: *mut TF_Status,
-    );
-}
-extern "C" {
-    pub fn TF_OperationGetAttrType(
-        oper: *mut TF_Operation,
-        attr_name: *const ::std::os::raw::c_char,
-        value: *mut TF_DataType,
-        status: *mut TF_Status,
-    );
-}
-extern "C" {
-    pub fn TF_OperationGetAttrTypeList(
-        oper: *mut TF_Operation,
-        attr_name: *const ::std::os::raw::c_char,
-        values: *mut TF_DataType,
-        max_values: ::std::os::raw::c_int,
-        status: *mut TF_Status,
-    );
-}
-extern "C" {
-    pub fn TF_OperationGetAttrShape(
-        oper: *mut TF_Operation,
-        attr_name: *const ::std::os::raw::c_char,
-        value: *mut i64,
-        num_dims: ::std::os::raw::c_int,
-        status: *mut TF_Status,
-    );
-}
-extern "C" {
-    pub fn TF_OperationGetAttrShapeList(
-        oper: *mut TF_Operation,
-        attr_name: *const ::std::os::raw::c_char,
-        dims: *mut *mut i64,
-        num_dims: *mut ::std::os::raw::c_int,
-        num_shapes: ::std::os::raw::c_int,
-        storage: *mut i64,
-        storage_size: ::std::os::raw::c_int,
-        status: *mut TF_Status,
-    );
-}
-extern "C" {
-    pub fn TF_OperationGetAttrTensorShapeProto(
-        oper: *mut TF_Operation,
-        attr_name: *const ::std::os::raw::c_char,
-        value: *mut TF_Buffer,
-        status: *mut TF_Status,
-    );
-}
-extern "C" {
-    pub fn TF_OperationGetAttrTensorShapeProtoList(
-        oper: *mut TF_Operation,
-        attr_name: *const ::std::os::raw::c_char,
-        values: *mut *mut TF_Buffer,
-        max_values: ::std::os::raw::c_int,
-        status: *mut TF_Status,
-    );
-}
-extern "C" {
-    pub fn TF_OperationGetAttrTensor(
-        oper: *mut TF_Operation,
-        attr_name: *const ::std::os::raw::c_char,
-        value: *mut *mut TF_Tensor,
-        status: *mut TF_Status,
-    );
-}
-extern "C" {
-    pub fn TF_OperationGetAttrTensorList(
-        oper: *mut TF_Operation,
-        attr_name: *const ::std::os::raw::c_char,
-        values: *mut *mut TF_Tensor,
-        max_values: ::std::os::raw::c_int,
-        status: *mut TF_Status,
-    );
-}
-extern "C" {
-    pub fn TF_OperationGetAttrValueProto(
-        oper: *mut TF_Operation,
-        attr_name: *const ::std::os::raw::c_char,
-        output_attr_value: *mut TF_Buffer,
-        status: *mut TF_Status,
-    );
-}
-extern "C" {
-    pub fn TF_GraphOperationByName(
-        graph: *mut TF_Graph,
-        oper_name: *const ::std::os::raw::c_char,
-    ) -> *mut TF_Operation;
-}
-extern "C" {
-    pub fn TF_GraphNextOperation(graph: *mut TF_Graph, pos: *mut usize) -> *mut TF_Operation;
-}
-extern "C" {
-    pub fn TF_GraphToGraphDef(
-        graph: *mut TF_Graph,
-        output_graph_def: *mut TF_Buffer,
-        status: *mut TF_Status,
-    );
-}
-extern "C" {
-    pub fn TF_GraphGetOpDef(
-        graph: *mut TF_Graph,
-        op_name: *const ::std::os::raw::c_char,
-        output_op_def: *mut TF_Buffer,
-        status: *mut TF_Status,
-    );
-}
-extern "C" {
-    pub fn TF_GraphVersions(
-        graph: *mut TF_Graph,
-        output_version_def: *mut TF_Buffer,
-        status: *mut TF_Status,
-    );
-}
-#[repr(C)]
-#[derive(Debug, Copy, Clone)]
-pub struct TF_ImportGraphDefOptions {
-    _unused: [u8; 0],
-}
-extern "C" {
-    pub fn TF_NewImportGraphDefOptions() -> *mut TF_ImportGraphDefOptions;
-}
-extern "C" {
-    pub fn TF_DeleteImportGraphDefOptions(opts: *mut TF_ImportGraphDefOptions);
-}
-extern "C" {
-    pub fn TF_ImportGraphDefOptionsSetPrefix(
-        opts: *mut TF_ImportGraphDefOptions,
-        prefix: *const ::std::os::raw::c_char,
-    );
-}
-extern "C" {
-    pub fn TF_ImportGraphDefOptionsSetDefaultDevice(
-        opts: *mut TF_ImportGraphDefOptions,
-        device: *const ::std::os::raw::c_char,
-    );
-}
-extern "C" {
-    pub fn TF_ImportGraphDefOptionsSetUniquifyNames(
-        opts: *mut TF_ImportGraphDefOptions,
-        uniquify_names: ::std::os::raw::c_uchar,
-    );
-}
-extern "C" {
-    pub fn TF_ImportGraphDefOptionsSetUniquifyPrefix(
-        opts: *mut TF_ImportGraphDefOptions,
-        uniquify_prefix: ::std::os::raw::c_uchar,
-    );
-}
-extern "C" {
-    pub fn TF_ImportGraphDefOptionsAddInputMapping(
-        opts: *mut TF_ImportGraphDefOptions,
-        src_name: *const ::std::os::raw::c_char,
-        src_index: ::std::os::raw::c_int,
-        dst: TF_Output,
-    );
-}
-extern "C" {
-    pub fn TF_ImportGraphDefOptionsRemapControlDependency(
-        opts: *mut TF_ImportGraphDefOptions,
-        src_name: *const ::std::os::raw::c_char,
-        dst: *mut TF_Operation,
-    );
-}
-extern "C" {
-    pub fn TF_ImportGraphDefOptionsAddControlDependency(
-        opts: *mut TF_ImportGraphDefOptions,
-        oper: *mut TF_Operation,
-    );
-}
-extern "C" {
-    pub fn TF_ImportGraphDefOptionsAddReturnOutput(
-        opts: *mut TF_ImportGraphDefOptions,
-        oper_name: *const ::std::os::raw::c_char,
-        index: ::std::os::raw::c_int,
-    );
-}
-extern "C" {
-    pub fn TF_ImportGraphDefOptionsNumReturnOutputs(
-        opts: *const TF_ImportGraphDefOptions,
-    ) -> ::std::os::raw::c_int;
-}
-extern "C" {
-    pub fn TF_ImportGraphDefOptionsAddReturnOperation(
-        opts: *mut TF_ImportGraphDefOptions,
-        oper_name: *const ::std::os::raw::c_char,
-    );
-}
-extern "C" {
-    pub fn TF_ImportGraphDefOptionsNumReturnOperations(
-        opts: *const TF_ImportGraphDefOptions,
-    ) -> ::std::os::raw::c_int;
-}
-#[repr(C)]
-#[derive(Debug, Copy, Clone)]
-pub struct TF_ImportGraphDefResults {
-    _unused: [u8; 0],
-}
-extern "C" {
-    pub fn TF_ImportGraphDefResultsReturnOutputs(
-        results: *mut TF_ImportGraphDefResults,
-        num_outputs: *mut ::std::os::raw::c_int,
-        outputs: *mut *mut TF_Output,
-    );
-}
-extern "C" {
-    pub fn TF_ImportGraphDefResultsReturnOperations(
-        results: *mut TF_ImportGraphDefResults,
-        num_opers: *mut ::std::os::raw::c_int,
-        opers: *mut *mut *mut TF_Operation,
-    );
-}
-extern "C" {
-    pub fn TF_ImportGraphDefResultsMissingUnusedInputMappings(
-        results: *mut TF_ImportGraphDefResults,
-        num_missing_unused_input_mappings: *mut ::std::os::raw::c_int,
-        src_names: *mut *mut *const ::std::os::raw::c_char,
-        src_indexes: *mut *mut ::std::os::raw::c_int,
-    );
-}
-extern "C" {
-    pub fn TF_DeleteImportGraphDefResults(results: *mut TF_ImportGraphDefResults);
-}
-extern "C" {
-    pub fn TF_GraphImportGraphDefWithResults(
-        graph: *mut TF_Graph,
-        graph_def: *const TF_Buffer,
-        options: *const TF_ImportGraphDefOptions,
-        status: *mut TF_Status,
-    ) -> *mut TF_ImportGraphDefResults;
-}
-extern "C" {
-    pub fn TF_GraphImportGraphDefWithReturnOutputs(
-        graph: *mut TF_Graph,
-        graph_def: *const TF_Buffer,
-        options: *const TF_ImportGraphDefOptions,
-        return_outputs: *mut TF_Output,
-        num_return_outputs: ::std::os::raw::c_int,
-        status: *mut TF_Status,
-    );
-}
-extern "C" {
-    pub fn TF_GraphImportGraphDef(
-        graph: *mut TF_Graph,
-        graph_def: *const TF_Buffer,
-        options: *const TF_ImportGraphDefOptions,
-        status: *mut TF_Status,
-    );
-}
-extern "C" {
-    pub fn TF_GraphCopyFunction(
-        g: *mut TF_Graph,
-        func: *const TF_Function,
-        grad: *const TF_Function,
-        status: *mut TF_Status,
-    );
-}
-extern "C" {
-    pub fn TF_GraphNumFunctions(g: *mut TF_Graph) -> ::std::os::raw::c_int;
-}
-extern "C" {
-    pub fn TF_GraphGetFunctions(
-        g: *mut TF_Graph,
-        funcs: *mut *mut TF_Function,
-        max_func: ::std::os::raw::c_int,
-        status: *mut TF_Status,
-    ) -> ::std::os::raw::c_int;
-}
-extern "C" {
-    pub fn TF_OperationToNodeDef(
-        oper: *mut TF_Operation,
-        output_node_def: *mut TF_Buffer,
-        status: *mut TF_Status,
-    );
-}
-#[repr(C)]
-#[derive(Debug, Copy, Clone)]
-pub struct TF_WhileParams {
-    pub ninputs: ::std::os::raw::c_int,
-    pub cond_graph: *mut TF_Graph,
-    pub cond_inputs: *const TF_Output,
-    pub cond_output: TF_Output,
-    pub body_graph: *mut TF_Graph,
-    pub body_inputs: *const TF_Output,
-    pub body_outputs: *mut TF_Output,
-    pub name: *const ::std::os::raw::c_char,
-}
-#[test]
-fn bindgen_test_layout_TF_WhileParams() {
-    assert_eq!(
-        ::std::mem::size_of::<TF_WhileParams>(),
-        72usize,
-        concat!("Size of: ", stringify!(TF_WhileParams))
-    );
-    assert_eq!(
-        ::std::mem::align_of::<TF_WhileParams>(),
-        8usize,
-        concat!("Alignment of ", stringify!(TF_WhileParams))
-    );
-    assert_eq!(
-        unsafe { &(*(::std::ptr::null::<TF_WhileParams>())).ninputs as *const _ as usize },
-        0usize,
-        concat!(
-            "Offset of field: ",
-            stringify!(TF_WhileParams),
-            "::",
-            stringify!(ninputs)
-        )
-    );
-    assert_eq!(
-        unsafe { &(*(::std::ptr::null::<TF_WhileParams>())).cond_graph as *const _ as usize },
-        8usize,
-        concat!(
-            "Offset of field: ",
-            stringify!(TF_WhileParams),
-            "::",
-            stringify!(cond_graph)
-        )
-    );
-    assert_eq!(
-        unsafe { &(*(::std::ptr::null::<TF_WhileParams>())).cond_inputs as *const _ as usize },
-        16usize,
-        concat!(
-            "Offset of field: ",
-            stringify!(TF_WhileParams),
-            "::",
-            stringify!(cond_inputs)
-        )
-    );
-    assert_eq!(
-        unsafe { &(*(::std::ptr::null::<TF_WhileParams>())).cond_output as *const _ as usize },
-        24usize,
-        concat!(
-            "Offset of field: ",
-            stringify!(TF_WhileParams),
-            "::",
-            stringify!(cond_output)
-        )
-    );
-    assert_eq!(
-        unsafe { &(*(::std::ptr::null::<TF_WhileParams>())).body_graph as *const _ as usize },
-        40usize,
-        concat!(
-            "Offset of field: ",
-            stringify!(TF_WhileParams),
-            "::",
-            stringify!(body_graph)
-        )
-    );
-    assert_eq!(
-        unsafe { &(*(::std::ptr::null::<TF_WhileParams>())).body_inputs as *const _ as usize },
-        48usize,
-        concat!(
-            "Offset of field: ",
-            stringify!(TF_WhileParams),
-            "::",
-            stringify!(body_inputs)
-        )
-    );
-    assert_eq!(
-        unsafe { &(*(::std::ptr::null::<TF_WhileParams>())).body_outputs as *const _ as usize },
-        56usize,
-        concat!(
-            "Offset of field: ",
-            stringify!(TF_WhileParams),
-            "::",
-            stringify!(body_outputs)
-        )
-    );
-    assert_eq!(
-        unsafe { &(*(::std::ptr::null::<TF_WhileParams>())).name as *const _ as usize },
-        64usize,
-        concat!(
-            "Offset of field: ",
-            stringify!(TF_WhileParams),
-            "::",
-            stringify!(name)
-        )
-    );
-}
-extern "C" {
-    pub fn TF_NewWhile(
-        g: *mut TF_Graph,
-        inputs: *mut TF_Output,
-        ninputs: ::std::os::raw::c_int,
-        status: *mut TF_Status,
-    ) -> TF_WhileParams;
-}
-extern "C" {
-    pub fn TF_FinishWhile(
-        params: *const TF_WhileParams,
-        status: *mut TF_Status,
-        outputs: *mut TF_Output,
-    );
-}
-extern "C" {
-    pub fn TF_AbortWhile(params: *const TF_WhileParams);
-}
-extern "C" {
-    pub fn TF_AddGradients(
-        g: *mut TF_Graph,
-        y: *mut TF_Output,
-        ny: ::std::os::raw::c_int,
-        x: *mut TF_Output,
-        nx: ::std::os::raw::c_int,
-        dx: *mut TF_Output,
-        status: *mut TF_Status,
-        dy: *mut TF_Output,
-    );
-}
-extern "C" {
-    pub fn TF_AddGradientsWithPrefix(
-        g: *mut TF_Graph,
-        prefix: *const ::std::os::raw::c_char,
-        y: *mut TF_Output,
-        ny: ::std::os::raw::c_int,
-        x: *mut TF_Output,
-        nx: ::std::os::raw::c_int,
-        dx: *mut TF_Output,
-        status: *mut TF_Status,
-        dy: *mut TF_Output,
-    );
-}
-extern "C" {
-    pub fn TF_GraphToFunction(
-        fn_body: *const TF_Graph,
-        fn_name: *const ::std::os::raw::c_char,
-        append_hash_to_fn_name: ::std::os::raw::c_uchar,
-        num_opers: ::std::os::raw::c_int,
-        opers: *const *const TF_Operation,
-        ninputs: ::std::os::raw::c_int,
-        inputs: *const TF_Output,
-        noutputs: ::std::os::raw::c_int,
-        outputs: *const TF_Output,
-        output_names: *const *const ::std::os::raw::c_char,
-        opts: *const TF_FunctionOptions,
-        description: *const ::std::os::raw::c_char,
-        status: *mut TF_Status,
-    ) -> *mut TF_Function;
-}
-extern "C" {
-    pub fn TF_GraphToFunctionWithControlOutputs(
-        fn_body: *const TF_Graph,
-        fn_name: *const ::std::os::raw::c_char,
-        append_hash_to_fn_name: ::std::os::raw::c_uchar,
-        num_opers: ::std::os::raw::c_int,
-        opers: *const *const TF_Operation,
-        ninputs: ::std::os::raw::c_int,
-        inputs: *const TF_Output,
-        noutputs: ::std::os::raw::c_int,
-        outputs: *const TF_Output,
-        output_names: *const *const ::std::os::raw::c_char,
-        ncontrol_outputs: ::std::os::raw::c_int,
-        control_outputs: *const *const TF_Operation,
-        control_output_names: *const *const ::std::os::raw::c_char,
-        opts: *const TF_FunctionOptions,
-        description: *const ::std::os::raw::c_char,
-        status: *mut TF_Status,
-    ) -> *mut TF_Function;
-}
-extern "C" {
-    pub fn TF_FunctionName(func: *mut TF_Function) -> *const ::std::os::raw::c_char;
-}
-extern "C" {
-    pub fn TF_FunctionToFunctionDef(
-        func: *mut TF_Function,
-        output_func_def: *mut TF_Buffer,
-        status: *mut TF_Status,
-    );
-}
-extern "C" {
-    pub fn TF_FunctionImportFunctionDef(
-        proto: *const ::std::os::raw::c_void,
-        proto_len: usize,
-        status: *mut TF_Status,
-    ) -> *mut TF_Function;
-}
-extern "C" {
-    pub fn TF_FunctionSetAttrValueProto(
-        func: *mut TF_Function,
-        attr_name: *const ::std::os::raw::c_char,
-        proto: *const ::std::os::raw::c_void,
-        proto_len: usize,
-        status: *mut TF_Status,
-    );
-}
-extern "C" {
-    pub fn TF_FunctionGetAttrValueProto(
-        func: *mut TF_Function,
-        attr_name: *const ::std::os::raw::c_char,
-        output_attr_value: *mut TF_Buffer,
-        status: *mut TF_Status,
-    );
-}
-extern "C" {
-    pub fn TF_DeleteFunction(func: *mut TF_Function);
-}
-extern "C" {
-    pub fn TF_TryEvaluateConstant(
-        graph: *mut TF_Graph,
-        output: TF_Output,
-        result: *mut *mut TF_Tensor,
-        status: *mut TF_Status,
-    ) -> ::std::os::raw::c_uchar;
-}
-#[repr(C)]
-#[derive(Debug, Copy, Clone)]
-pub struct TF_Session {
-    _unused: [u8; 0],
-}
-extern "C" {
-    pub fn TF_NewSession(
-        graph: *mut TF_Graph,
-        opts: *const TF_SessionOptions,
-        status: *mut TF_Status,
-    ) -> *mut TF_Session;
-}
-extern "C" {
-    pub fn TF_LoadSessionFromSavedModel(
-        session_options: *const TF_SessionOptions,
-        run_options: *const TF_Buffer,
-        export_dir: *const ::std::os::raw::c_char,
-        tags: *const *const ::std::os::raw::c_char,
-        tags_len: ::std::os::raw::c_int,
-        graph: *mut TF_Graph,
-        meta_graph_def: *mut TF_Buffer,
-        status: *mut TF_Status,
-    ) -> *mut TF_Session;
-}
-extern "C" {
-    pub fn TF_CloseSession(arg1: *mut TF_Session, status: *mut TF_Status);
-}
-extern "C" {
-    pub fn TF_DeleteSession(arg1: *mut TF_Session, status: *mut TF_Status);
-}
-extern "C" {
-    pub fn TF_SessionRun(
-        session: *mut TF_Session,
-        run_options: *const TF_Buffer,
-        inputs: *const TF_Output,
-        input_values: *const *mut TF_Tensor,
-        ninputs: ::std::os::raw::c_int,
-        outputs: *const TF_Output,
-        output_values: *mut *mut TF_Tensor,
-        noutputs: ::std::os::raw::c_int,
-        target_opers: *const *const TF_Operation,
-        ntargets: ::std::os::raw::c_int,
-        run_metadata: *mut TF_Buffer,
-        arg1: *mut TF_Status,
-    );
-}
-extern "C" {
-    pub fn TF_SessionPRunSetup(
-        arg1: *mut TF_Session,
-        inputs: *const TF_Output,
-        ninputs: ::std::os::raw::c_int,
-        outputs: *const TF_Output,
-        noutputs: ::std::os::raw::c_int,
-        target_opers: *const *const TF_Operation,
-        ntargets: ::std::os::raw::c_int,
-        handle: *mut *const ::std::os::raw::c_char,
-        arg2: *mut TF_Status,
-    );
-}
-extern "C" {
-    pub fn TF_SessionPRun(
-        arg1: *mut TF_Session,
-        handle: *const ::std::os::raw::c_char,
-        inputs: *const TF_Output,
-        input_values: *const *mut TF_Tensor,
-        ninputs: ::std::os::raw::c_int,
-        outputs: *const TF_Output,
-        output_values: *mut *mut TF_Tensor,
-        noutputs: ::std::os::raw::c_int,
-        target_opers: *const *const TF_Operation,
-        ntargets: ::std::os::raw::c_int,
-        arg2: *mut TF_Status,
-    );
-}
-extern "C" {
-    pub fn TF_DeletePRunHandle(handle: *const ::std::os::raw::c_char);
-}
-#[repr(C)]
-#[derive(Debug, Copy, Clone)]
-pub struct TF_DeprecatedSession {
-    _unused: [u8; 0],
-}
-extern "C" {
-    pub fn TF_NewDeprecatedSession(
-        arg1: *const TF_SessionOptions,
-        status: *mut TF_Status,
-    ) -> *mut TF_DeprecatedSession;
-}
-extern "C" {
-    pub fn TF_CloseDeprecatedSession(arg1: *mut TF_DeprecatedSession, status: *mut TF_Status);
-}
-extern "C" {
-    pub fn TF_DeleteDeprecatedSession(arg1: *mut TF_DeprecatedSession, status: *mut TF_Status);
-}
-extern "C" {
-    pub fn TF_Reset(
-        opt: *const TF_SessionOptions,
-        containers: *mut *const ::std::os::raw::c_char,
-        ncontainers: ::std::os::raw::c_int,
-        status: *mut TF_Status,
-    );
-}
-extern "C" {
-    pub fn TF_ExtendGraph(
-        arg1: *mut TF_DeprecatedSession,
-        proto: *const ::std::os::raw::c_void,
-        proto_len: usize,
-        arg2: *mut TF_Status,
-    );
-}
-extern "C" {
-    pub fn TF_Run(
-        arg1: *mut TF_DeprecatedSession,
-        run_options: *const TF_Buffer,
-        input_names: *mut *const ::std::os::raw::c_char,
-        inputs: *mut *mut TF_Tensor,
-        ninputs: ::std::os::raw::c_int,
-        output_names: *mut *const ::std::os::raw::c_char,
-        outputs: *mut *mut TF_Tensor,
-        noutputs: ::std::os::raw::c_int,
-        target_oper_names: *mut *const ::std::os::raw::c_char,
-        ntargets: ::std::os::raw::c_int,
-        run_metadata: *mut TF_Buffer,
-        arg2: *mut TF_Status,
-    );
-}
-extern "C" {
-    pub fn TF_PRunSetup(
-        arg1: *mut TF_DeprecatedSession,
-        input_names: *mut *const ::std::os::raw::c_char,
-        ninputs: ::std::os::raw::c_int,
-        output_names: *mut *const ::std::os::raw::c_char,
-        noutputs: ::std::os::raw::c_int,
-        target_oper_names: *mut *const ::std::os::raw::c_char,
-        ntargets: ::std::os::raw::c_int,
-        handle: *mut *const ::std::os::raw::c_char,
-        arg2: *mut TF_Status,
-    );
-}
-extern "C" {
-    pub fn TF_PRun(
-        arg1: *mut TF_DeprecatedSession,
-        handle: *const ::std::os::raw::c_char,
-        input_names: *mut *const ::std::os::raw::c_char,
-        inputs: *mut *mut TF_Tensor,
-        ninputs: ::std::os::raw::c_int,
-        output_names: *mut *const ::std::os::raw::c_char,
-        outputs: *mut *mut TF_Tensor,
-        noutputs: ::std::os::raw::c_int,
-        target_oper_names: *mut *const ::std::os::raw::c_char,
-        ntargets: ::std::os::raw::c_int,
-        arg2: *mut TF_Status,
-    );
-}
-#[repr(C)]
-#[derive(Debug, Copy, Clone)]
-pub struct TF_DeviceList {
-    _unused: [u8; 0],
-}
-extern "C" {
-    pub fn TF_SessionListDevices(
-        session: *mut TF_Session,
-        status: *mut TF_Status,
-    ) -> *mut TF_DeviceList;
-}
-extern "C" {
-    pub fn TF_DeprecatedSessionListDevices(
-        session: *mut TF_DeprecatedSession,
-        status: *mut TF_Status,
-    ) -> *mut TF_DeviceList;
-}
-extern "C" {
-    pub fn TF_DeleteDeviceList(list: *mut TF_DeviceList);
-}
-extern "C" {
-    pub fn TF_DeviceListCount(list: *const TF_DeviceList) -> ::std::os::raw::c_int;
-}
-extern "C" {
-    pub fn TF_DeviceListName(
-        list: *const TF_DeviceList,
-        index: ::std::os::raw::c_int,
-        status: *mut TF_Status,
-    ) -> *const ::std::os::raw::c_char;
-}
-extern "C" {
-    pub fn TF_DeviceListType(
-        list: *const TF_DeviceList,
-        index: ::std::os::raw::c_int,
-        status: *mut TF_Status,
-    ) -> *const ::std::os::raw::c_char;
-}
-extern "C" {
-    pub fn TF_DeviceListMemoryBytes(
-        list: *const TF_DeviceList,
-        index: ::std::os::raw::c_int,
-        status: *mut TF_Status,
-    ) -> i64;
-}
-extern "C" {
-    pub fn TF_DeviceListIncarnation(
-        list: *const TF_DeviceList,
-        index: ::std::os::raw::c_int,
-        status: *mut TF_Status,
-    ) -> u64;
-}
-#[repr(C)]
-#[derive(Debug, Copy, Clone)]
-pub struct TF_Library {
-    _unused: [u8; 0],
-}
-extern "C" {
-    pub fn TF_LoadLibrary(
-        library_filename: *const ::std::os::raw::c_char,
-        status: *mut TF_Status,
-    ) -> *mut TF_Library;
-}
-extern "C" {
-    pub fn TF_GetOpList(lib_handle: *mut TF_Library) -> TF_Buffer;
-}
-extern "C" {
-    pub fn TF_DeleteLibraryHandle(lib_handle: *mut TF_Library);
-}
-extern "C" {
-    pub fn TF_GetAllOpList() -> *mut TF_Buffer;
-}
-#[repr(C)]
-#[derive(Debug, Copy, Clone)]
-pub struct TF_ApiDefMap {
-    _unused: [u8; 0],
-}
-extern "C" {
-    pub fn TF_NewApiDefMap(
-        op_list_buffer: *mut TF_Buffer,
-        status: *mut TF_Status,
-    ) -> *mut TF_ApiDefMap;
-}
-extern "C" {
-    pub fn TF_DeleteApiDefMap(apimap: *mut TF_ApiDefMap);
-}
-extern "C" {
-    pub fn TF_ApiDefMapPut(
-        api_def_map: *mut TF_ApiDefMap,
-        text: *const ::std::os::raw::c_char,
-        text_len: usize,
-        status: *mut TF_Status,
-    );
-}
-extern "C" {
-    pub fn TF_ApiDefMapGet(
-        api_def_map: *mut TF_ApiDefMap,
-        name: *const ::std::os::raw::c_char,
-        name_len: usize,
-        status: *mut TF_Status,
-    ) -> *mut TF_Buffer;
-}
-extern "C" {
-    pub fn TF_GetAllRegisteredKernels(status: *mut TF_Status) -> *mut TF_Buffer;
-}
-extern "C" {
-    pub fn TF_GetRegisteredKernelsForOp(
-        name: *const ::std::os::raw::c_char,
-        status: *mut TF_Status,
-    ) -> *mut TF_Buffer;
-}
-extern "C" {
-    pub fn TF_UpdateEdge(
-        graph: *mut TF_Graph,
-        new_src: TF_Output,
-        dst: TF_Input,
-        status: *mut TF_Status,
-    );
-}
-#[repr(C)]
-#[derive(Debug, Copy, Clone)]
-pub struct TF_Server {
-    _unused: [u8; 0],
-}
-extern "C" {
-    pub fn TF_NewServer(
-        proto: *const ::std::os::raw::c_void,
-        proto_len: usize,
-        status: *mut TF_Status,
-    ) -> *mut TF_Server;
-}
-extern "C" {
-    pub fn TF_ServerStart(server: *mut TF_Server, status: *mut TF_Status);
-}
-extern "C" {
-    pub fn TF_ServerStop(server: *mut TF_Server, status: *mut TF_Status);
-}
-extern "C" {
-    pub fn TF_ServerJoin(server: *mut TF_Server, status: *mut TF_Status);
-}
-extern "C" {
-    pub fn TF_ServerTarget(server: *mut TF_Server) -> *const ::std::os::raw::c_char;
-}
-extern "C" {
-    pub fn TF_DeleteServer(server: *mut TF_Server);
-}
-extern "C" {
-    pub fn TF_RegisterLogListener(
-        listener: ::std::option::Option<unsafe extern "C" fn(arg1: *const ::std::os::raw::c_char)>,
-    );
-}
-extern "C" {
-    pub fn TF_RegisterFilesystemPlugin(
-        plugin_filename: *const ::std::os::raw::c_char,
-        status: *mut TF_Status,
-    );
-}
-#[repr(C)]
-#[derive(Debug, Copy, Clone)]
-pub struct __locale_data {
-    pub _address: u8,
-}
