+++ tensorflow-sys/src/eager/mod.rs
@@ -1,0 +1,5 @@
+use crate::{
+    TF_AttrType, TF_Buffer, TF_DataType, TF_DeviceList, TF_Function, TF_Status, TF_Tensor,
+};
+
+include!("c_api.rs");
